Bayesian graphical clustering

Perform Bayesian graphical clustering of the differential analysis results using repfdr

Usage

bayesian_graphical_clustering(
  zscores,
  min_analyte_posterior_thr = 0.5,
  min_prior_for_config = 0.001,
  df = 20,
  naive_edge_sets = TRUE
)

Arguments

zscores

A numeric matrix. Rows are analytes, columns are conditions (e.g., male week 1). Entries are z-scores.

min_analyte_posterior_thr

A number. The minimal value of sum of posteriors for including an analyte.

min_prior_for_config

A number. The minimal prior probability on the configuration priors.

df

Integer. Parameter for repfdr::ztobins(). Degrees of freedom for fitting the estimated density f(z).

naive_edge_sets

A logical. TRUE: edge sets are extracted by taking simple intersection of nodes. FALSE: edge sets are extracted from the repfdr solution.

Value

A named list with three items:

node_sets

named list of node sets

edge_sets

named list of edge sets

repfdr_results

repfdr EM solution

Details

This function is highly specialized for the MoTrPAC data. It assumes that the input zscore matrix has eight columns. Four columns for male z-scores (weeks 1, 2, 4, 8), and four for females. No NA values in the input matrix. The function first runs the Bayesian clustering using repfdr (see repfdr_wrapper()) and then extracts the node and edge sets of the graphical solution.

A node is a time-point specific state. It can represent null features (z-score around zero, no effect), up-regulated features (positive z-score), or down regulated features (negative z-scores). An edge represents interaction among two adjacent time point. For example, it can represent features that are up-regulated in week 1 in both males and females, and null in both male and females by week 2. min_analyte_posterior_thr is used to make sure that features without any reasonable fit in the clustering solution will be discarded, whereas min_prior_for_config removes clusters with extremely low prior (i.e., do not have any associated features).

Formally, let zscores be the timewise nxm matrix of z-scores over n analytes and m conditions, where m is even and has the same time points for males and females. Each z-score corresponds to the effect of training compared to untrained baseline. We assume that z-scores have a latent configuration in -1,0,1, with -1 denotes down-regulation, 0 denotes null (no change), and 1 denotes up-regulation. Let h denote the latent configuration of an analyte over all m conditions. We use the expectation-maximization (EM) process of the repfdr algorithm (Heller and Yekutieli, 2014) to estimate the prior π(h) and posterior Pr(h|z_i), for every analyte i. Once the algorithm converges, we discard configurations h with π(h) < min_analyte_posterior_thr and normalize Pr(h|zi) to sum to 1 (i.e., all posteriors given the same zi). The new posteriors that can be interpreted as a soft clustering solution, where the greater the value is, the more likely it is for analyte i to participate in cluster h.

Given the posteriors Pr(h|zi), we assign analytes to nine possible states in each time point (male or females x time points 1w, 2w,4w, and 8w). Analyte i belongs to a state if the sum of posteriors that are consistent with that state is > min_analyte_posterior_thr. For every pair of states from adjacent time points j and j+1 we define their edge set as the set of analytes whose differential expression pattern is consistent with the two nodes.. Thus, the node and edge sets explained above together define a tree structure that represent different differential patterns over sex and time.

Naming format for the output: Node set names: {week}_F{differential regulation status}_M{differential regulation status}. Example 1: 1w_F1_M0 means up-regulation in females week 1 and null (zero effect) in males in week 1. Example 2: 1w_F1_M-1 means up-regulation in females week 1 and down-regulation in males in week 1. The edge set object is a named list of string vectors. The name of an edge is [node_id]---[node_id].

Examples

### Example 1: Simulate data with a single cluster
zcolnames = c(
  paste("female",c("1w","2w","4w","8w"),sep="_"),
  paste("male",c("1w","2w","4w","8w"),sep="_")
)
zscores = matrix(rnorm(80000), ncol=8, dimnames = list(1:10000,zcolnames))
# now add a cluster with a strong signal 
zscores[1:500,1:4] = zscores[1:500,1:4] + 5 

# Run the clustering solution wrapper.
# When the data are "clean" (e.g., a mixture of two gaussians), we do not
# need a high df in the two-groups model estimation
# (default is 20, consider at least 10 when analyzing real data)
clustering_sol = bayesian_graphical_clustering(zscores, df=5)
#> Warning: In column  2   female_2w ,f(z) misfit =  1.9 . Rerun with increased df
#> Warning: In column 7 male_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 male_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> Converged after 1 EM iterations.
#> Warning: In column  1   female_1w ,f(z) misfit =  5.8 . Rerun with increased df
#> Warning: In column  2   female_2w ,f(z) misfit =  7.2 . Rerun with increased df
#> Warning: In column  3   female_4w ,f(z) misfit =  5.5 . Rerun with increased df
#> Warning: In column  4   female_8w ,f(z) misfit =  6.8 . Rerun with increased df
#> Warning: In column 7 male_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 male_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> Converged after 1 EM iterations.
#> Warning: In column  1   female_1w ,f(z) misfit =  15 . Rerun with increased df
#> Warning: In column  2   female_2w ,f(z) misfit =  7.6 . Rerun with increased df
#> Warning: In column  3   female_4w ,f(z) misfit =  14.4 . Rerun with increased df
#> Warning: In column  4   female_8w ,f(z) misfit =  9 . Rerun with increased df
#> Warning: In column 5 male_1w the estimated fraction of nulls is 1.
#> Warning: In column 6 male_2w the estimated fraction of nulls is 1.
#> Warning: In column 7 male_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 male_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> EM iteration 3
#> EM iteration 4
#> EM iteration 5
#> EM iteration 6
#> EM iteration 7
#> EM iteration 8
#> EM iteration 9
#> EM iteration 10
#> EM iteration 11
#> EM iteration 12
#> EM iteration 13
#> EM iteration 14
#> EM iteration 15
#> EM iteration 16
#> EM iteration 17
#> EM iteration 18
#> EM iteration 19
#> EM iteration 20
#> EM iteration 21
#> EM iteration 22
#> EM iteration 23
#> EM iteration 24
#> EM iteration 25
#> EM iteration 26
#> EM iteration 27
#> EM iteration 28
#> EM iteration 29
#> EM iteration 30
#> EM iteration 31
#> EM iteration 32
#> EM iteration 33
#> EM iteration 34
#> EM iteration 35
#> EM iteration 36
#> EM iteration 37
#> EM iteration 38
#> EM iteration 39
#> EM iteration 40
#> EM iteration 41
#> EM iteration 42
#> EM iteration 43
#> EM iteration 44
#> EM iteration 45
#> EM iteration 46
#> EM iteration 47
#> EM iteration 48
#> EM iteration 49
#> EM iteration 50
#> EM iteration 51
#> EM iteration 52
#> EM iteration 53
#> EM iteration 54
#> EM iteration 55
#> EM iteration 56
#> EM iteration 57
#> EM iteration 58
#> EM iteration 59
#> EM iteration 60
#> EM iteration 61
#> EM iteration 62
#> EM iteration 63
#> EM iteration 64
#> EM iteration 65
#> EM iteration 66
#> EM iteration 67
#> EM iteration 68
#> EM iteration 69
#> EM iteration 70
#> EM iteration 71
#> EM iteration 72
#> EM iteration 73
#> EM iteration 74
#> EM iteration 75
#> EM iteration 76
#> EM iteration 77
#> EM iteration 78
#> EM iteration 79
#> EM iteration 80
#> EM iteration 81
#> EM iteration 82
#> EM iteration 83
#> EM iteration 84
#> EM iteration 85
#> EM iteration 86
#> EM iteration 87
#> EM iteration 88
#> EM iteration 89
#> EM iteration 90
#> EM iteration 91
#> EM iteration 92
#> EM iteration 93
#> EM iteration 94
#> EM iteration 95
#> EM iteration 96
#> EM iteration 97
#> EM iteration 98
#> Converged after 97 EM iterations.
# check if the clustering solution correctly assigns the first 500 rows 
# (with high prob) to the right nodes
length(intersect(1:500,clustering_sol$node_sets$`1w_F1_M0`))/500 > 0.95
#> [1] TRUE
length(intersect(1:500,clustering_sol$node_sets$`2w_F1_M0`))/500 > 0.95
#> [1] TRUE
length(intersect(1:500,clustering_sol$node_sets$`4w_F1_M0`))/500 > 0.95
#> [1] TRUE
length(intersect(1:500,clustering_sol$node_sets$`8w_F1_M0`))/500 > 0.95
#> [1] TRUE

# examine the node set sizes
sapply(clustering_sol$node_sets,length)
#>  1w_F0_M0  1w_F1_M0  2w_F0_M0  2w_F1_M0  4w_F0_M0  4w_F1_M0 8w_F-1_M0  8w_F0_M0 
#>      9493       505      9493       505      9493       505         2      9491 
#>  8w_F1_M0        0w 
#>       505     10000 
# examine the edge set sizes
sapply(clustering_sol$edge_sets,length)
#>        0w---1w_F0_M0        0w---1w_F1_M0  1w_F0_M0---2w_F0_M0 
#>                 9493                  505                 9493 
#>  1w_F1_M0---2w_F0_M0  1w_F0_M0---2w_F1_M0  1w_F1_M0---2w_F1_M0 
#>                    0                    0                  505 
#>  2w_F0_M0---4w_F0_M0  2w_F1_M0---4w_F0_M0  2w_F0_M0---4w_F1_M0 
#>                 9493                    0                    0 
#>  2w_F1_M0---4w_F1_M0 4w_F0_M0---8w_F-1_M0 4w_F1_M0---8w_F-1_M0 
#>                  505                    2                    0 
#>  4w_F0_M0---8w_F0_M0  4w_F1_M0---8w_F0_M0  4w_F0_M0---8w_F1_M0 
#>                 9491                    0                    0 
#>  4w_F1_M0---8w_F1_M0 
#>                  505 

# extract the top full trajectories in the data
# these should be the clusters with at least 10 features
min_cluster_size = 10
get_trajectory_sizes_from_edge_sets(clustering_sol$edge_sets,min_size = min_cluster_size)
#>   X1       X2       X3       X4       X5   X6
#> 1 0w 1w_F0_M0 2w_F0_M0 4w_F0_M0 8w_F0_M0 9491
#> 2 0w 1w_F1_M0 2w_F1_M0 4w_F1_M0 8w_F1_M0  505

if (FALSE) {
### Example 2: real data
data(REPFDR_INPUTS, package="MotrpacRatTraining6moData")
zscores = REPFDR_INPUTS$zs_smoothed
rat_data_clustering_sol = bayesian_graphical_clustering(zscores)
# extract the largest trajectories
get_trajectory_sizes_from_edge_sets(rat_data_clustering_sol$edge_sets, min_size = 50)
# plot the top trajectories of the muscle tissues, color edges by tissue
get_tree_plot_for_tissue(tissues = c("SKM-GN","HEART","SKM-VL"), 
                         omes = "TRNSCRPT",
                         node_sets = rat_data_clustering_sol$node_sets,
                         edge_sets = rat_data_clustering_sol$edge_sets,
                         min_size = 20,
                         parallel_edges_by_tissue = TRUE,
                         max_trajectories = 3)
}