Combine differential analysis results

Combine differential analysis results from the specified tissues and omes(s)/assay(s). If no tissues or omes are specified, all differential analysis results are returned.

Usage

combine_da_results(
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  metareg = TRUE,
  include_epigen = FALSE,
  scratchdir = "."
)

Arguments

tissues

optional character vector of tissue abbreviations, one of MotrpacRatTraining6moData::TISSUE_ABBREV.

assays

optional character vector of assay abbreviations, one of MotrpacRatTraining6moData::ASSAY_ABBREV

metareg

bool, whether to use the meta-regression results for METAB. Use the upstream differential analysis results in FALSE. TRUE by default.

include_epigen

bool, whether to include the full ATAC or METHYL differential analysis results from Google Cloud Storage. Only relevant if assays includes "ATAC" or "METHYL". FALSE by default.

scratchdir

character, local directory in which to download data from the web. Current working directory by default. Only relevant if assays includes "ATAC" or "METHYL".

Value

data frame. Depending on the specified assays, some of these columns may not be included:

feature

character, unique feature identifier in the format 'MotrpacRatTraining6moData::ASSAY_ABBREV;MotrpacRatTraining6moData::TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See MotrpacRatTraining6moData::REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

covariates

character, comma-separated list of adjustment variables or NA

numNAs

integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

dataset

character, immune panel, metabolomics platform, or ATAC-seq dataset name

site

character, Chemical Analysis Site (CAS) name. METAB only

is_targeted

logical, is this a targeted platform? METAB only

metabolite_refmet

character, RefMet name of metabolite. METAB only

cv

double, feature coefficient of variation in the dataset. METAB only

metabolite

character, name of metabolite as appears in the CAS's data. METAB only

control_cv

double, feature coefficient of variation in the dataset. METAB only

mz

double, mass over charge. METAB only

rt

double, retention time. METAB only

neutral_mass

double, neutral mass. METAB only

meta_reg_het_p

double, for metabolites with multiple measurements, the meta-regression heterogeneity p-value, where a smaller p-value indicates more disagreement between platforms. One value per feature (not per training group). METAB only

meta_reg_pvalue

double, for metabolites with multiple measurements, the meta-regression p-value. One value per feature (not per training group). METAB only

shrunk_logFC

double, log fold-change with shrinkage applied

shrunk_logFC_se

double, standard error of the shrunken log fold-change

zscore

double, z statistic

removed_samples

character, comma-separated list of vial labels excluded from differential analysis or NA

comparison_average_intensity_se

double, standard error of 'comparison_average_intensity'

reference_average_intensity_se

double, standard error of 'comparison_average_intensity'

Chr

integer, chromosome. METHYL only

Locus

character, base pair range of feature. METHYL only

EntrezID

character, Entrez ID of closest gene. METHYL only

Symbol

character, gene symbol of closest gene. METHYL only

Examples

# Return all non-epigenetic differential analysis results, 
# including meta-regression results for metabolomics
combine_da_results()
#> Warning: 'include_epigen' is FALSE. Excluding ATAC and METHYL results.
#> ACETYL_HEART_DA
#> ACETYL_LIVER_DA
#> IMMUNO_ADRNL_DA
#> IMMUNO_BAT_DA
#> IMMUNO_COLON_DA
#> IMMUNO_CORTEX_DA
#> IMMUNO_HEART_DA
#> IMMUNO_HIPPOC_DA
#> IMMUNO_KIDNEY_DA
#> IMMUNO_LIVER_DA
#> IMMUNO_LUNG_DA
#> IMMUNO_OVARY_DA
#> IMMUNO_PLASMA_DA
#> IMMUNO_SKMGN_DA
#> IMMUNO_SKMVL_DA
#> IMMUNO_SMLINT_DA
#> IMMUNO_SPLEEN_DA
#> IMMUNO_TESTES_DA
#> IMMUNO_WATSC_DA
#> METAB_ADRNL_DA_METAREG
#> METAB_BAT_DA_METAREG
#> METAB_COLON_DA_METAREG
#> METAB_CORTEX_DA_METAREG
#> METAB_HEART_DA_METAREG
#> METAB_HIPPOC_DA_METAREG
#> METAB_HYPOTH_DA_METAREG
#> METAB_KIDNEY_DA_METAREG
#> METAB_LIVER_DA_METAREG
#> METAB_LUNG_DA_METAREG
#> METAB_OVARY_DA_METAREG
#> METAB_PLASMA_DA_METAREG
#> METAB_SKMGN_DA_METAREG
#> METAB_SKMVL_DA_METAREG
#> METAB_SMLINT_DA_METAREG
#> METAB_SPLEEN_DA_METAREG
#> METAB_TESTES_DA_METAREG
#> METAB_VENACV_DA_METAREG
#> METAB_WATSC_DA_METAREG
#> PHOSPHO_CORTEX_DA
#> PHOSPHO_HEART_DA
#> PHOSPHO_KIDNEY_DA
#> PHOSPHO_LIVER_DA
#> PHOSPHO_LUNG_DA
#> PHOSPHO_SKMGN_DA
#> PHOSPHO_WATSC_DA
#> PROT_CORTEX_DA
#> PROT_HEART_DA
#> PROT_KIDNEY_DA
#> PROT_LIVER_DA
#> PROT_LUNG_DA
#> PROT_SKMGN_DA
#> PROT_WATSC_DA
#> TRNSCRPT_ADRNL_DA
#> TRNSCRPT_BAT_DA
#> TRNSCRPT_BLOOD_DA
#> TRNSCRPT_COLON_DA
#> TRNSCRPT_CORTEX_DA
#> TRNSCRPT_HEART_DA
#> TRNSCRPT_HIPPOC_DA
#> TRNSCRPT_HYPOTH_DA
#> TRNSCRPT_KIDNEY_DA
#> TRNSCRPT_LIVER_DA
#> TRNSCRPT_LUNG_DA
#> TRNSCRPT_OVARY_DA
#> TRNSCRPT_SKMGN_DA
#> TRNSCRPT_SKMVL_DA
#> TRNSCRPT_SMLINT_DA
#> TRNSCRPT_SPLEEN_DA
#> TRNSCRPT_TESTES_DA
#> TRNSCRPT_VENACV_DA
#> TRNSCRPT_WATSC_DA
#> UBIQ_HEART_DA
#> UBIQ_LIVER_DA
#>          feature  assay                assay_code tissue tissue_code
#>       1:    <NA> ACETYL                   prot-ac  HEART   t58-heart
#>       2:    <NA> ACETYL                   prot-ac  HEART   t58-heart
#>       3:    <NA> ACETYL                   prot-ac  HEART   t58-heart
#>       4:    <NA> ACETYL                   prot-ac  HEART   t58-heart
#>       5:    <NA> ACETYL                   prot-ac  HEART   t58-heart
#>      ---                                                            
#> 5279132:    <NA>   UBIQ prot-ub-protein-corrected  LIVER   t68-liver
#> 5279133:    <NA>   UBIQ prot-ub-protein-corrected  LIVER   t68-liver
#> 5279134:    <NA>   UBIQ prot-ub-protein-corrected  LIVER   t68-liver
#> 5279135:    <NA>   UBIQ prot-ub-protein-corrected  LIVER   t68-liver
#> 5279136:    <NA>   UBIQ prot-ub-protein-corrected  LIVER   t68-liver
#>                 feature_ID    sex comparison_group   p_value adj_p_value
#>       1:  AP_004896.1_K46k   male               1w 0.8916202           1
#>       2:  AP_004896.1_K46k   male               2w 0.6971072           1
#>       3:  AP_004896.1_K46k   male               4w 0.3338958           1
#>       4:  AP_004896.1_K46k   male               8w 0.3615084           1
#>       5:  AP_004896.1_K46k female               1w 0.4877300           1
#>      ---                                                                
#> 5279132: XP_221997.5_K254k   male               8w        NA          NA
#> 5279133: XP_221997.5_K254k female               1w        NA          NA
#> 5279134: XP_221997.5_K254k female               2w        NA          NA
#> 5279135: XP_221997.5_K254k female               4w        NA          NA
#> 5279136: XP_221997.5_K254k female               8w        NA          NA
#>                logFC   logFC_se     tscore covariates numNAs
#>       1: -0.01798774 0.13051973 -0.1378163       <NA>      0
#>       2: -0.04267032 0.10822124 -0.3942879       <NA>      0
#>       3:  0.11163865 0.11303340  0.9876608       <NA>      0
#>       4: -0.10530057 0.11303340 -0.9315881       <NA>      0
#>       5:  0.04801416 0.06827064  0.7032916       <NA>      0
#>      ---                                                    
#> 5279132:          NA         NA         NA       <NA>     12
#> 5279133:          NA         NA         NA       <NA>     12
#> 5279134:          NA         NA         NA       <NA>     12
#> 5279135:          NA         NA         NA       <NA>     12
#> 5279136:          NA         NA         NA       <NA>     12
#>          comparison_average_intensity reference_average_intensity selection_fdr
#>       1:                  -0.16397535                -0.145987604     0.2435669
#>       2:                  -0.18865793                -0.145987604     0.2435669
#>       3:                  -0.03434895                -0.145987604     0.2435669
#>       4:                  -0.25128818                -0.145987604     0.2435669
#>       5:                   0.04193526                -0.006078899     0.2435669
#>      ---                                                                       
#> 5279132:                           NA                          NA     0.9447052
#> 5279133:                           NA                          NA     0.9447052
#> 5279134:                           NA                          NA     0.9447052
#> 5279135:                           NA                          NA     0.9447052
#> 5279136:                           NA                          NA     0.9447052
#>          dataset site is_targeted zscore metabolite_refmet cv metabolite
#>       1:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>       2:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>       3:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>       4:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>       5:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>      ---                                                                
#> 5279132:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#> 5279133:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#> 5279134:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#> 5279135:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#> 5279136:    <NA> <NA>          NA     NA              <NA> NA       <NA>
#>          control_cv mz rt neutral_mass meta_reg_het_p meta_reg_pvalue
#>       1:         NA NA NA           NA             NA              NA
#>       2:         NA NA NA           NA             NA              NA
#>       3:         NA NA NA           NA             NA              NA
#>       4:         NA NA NA           NA             NA              NA
#>       5:         NA NA NA           NA             NA              NA
#>      ---                                                             
#> 5279132:         NA NA NA           NA             NA              NA
#> 5279133:         NA NA NA           NA             NA              NA
#> 5279134:         NA NA NA           NA             NA              NA
#> 5279135:         NA NA NA           NA             NA              NA
#> 5279136:         NA NA NA           NA             NA              NA
#>          shrunk_logFC shrunk_logFC_se removed_samples
#>       1:           NA              NA            <NA>
#>       2:           NA              NA            <NA>
#>       3:           NA              NA            <NA>
#>       4:           NA              NA            <NA>
#>       5:           NA              NA            <NA>
#>      ---                                             
#> 5279132:           NA              NA            <NA>
#> 5279133:           NA              NA            <NA>
#> 5279134:           NA              NA            <NA>
#> 5279135:           NA              NA            <NA>
#> 5279136:           NA              NA            <NA>
#>          comparison_average_intensity_se reference_average_intensity_se
#>       1:                              NA                             NA
#>       2:                              NA                             NA
#>       3:                              NA                             NA
#>       4:                              NA                             NA
#>       5:                              NA                             NA
#>      ---                                                               
#> 5279132:                              NA                             NA
#> 5279133:                              NA                             NA
#> 5279134:                              NA                             NA
#> 5279135:                              NA                             NA
#> 5279136:                              NA                             NA

# Return all global proteomics differential analysis results
combine_da_results(assays="PROT")
#> PROT_CORTEX_DA
#> PROT_HEART_DA
#> PROT_KIDNEY_DA
#> PROT_LIVER_DA
#> PROT_LUNG_DA
#> PROT_SKMGN_DA
#> PROT_WATSC_DA
#>         feature assay assay_code tissue       tissue_code  feature_ID    sex
#>      1:    <NA>  PROT    prot-pr CORTEX        t53-cortex AP_004892.1   male
#>      2:    <NA>  PROT    prot-pr CORTEX        t53-cortex AP_004892.1   male
#>      3:    <NA>  PROT    prot-pr CORTEX        t53-cortex AP_004892.1   male
#>      4:    <NA>  PROT    prot-pr CORTEX        t53-cortex AP_004892.1   male
#>      5:    <NA>  PROT    prot-pr CORTEX        t53-cortex AP_004892.1 female
#>     ---                                                                     
#> 534196:    <NA>  PROT    prot-pr WAT-SC t70-white-adipose YP_665639.1   male
#> 534197:    <NA>  PROT    prot-pr WAT-SC t70-white-adipose YP_665639.1 female
#> 534198:    <NA>  PROT    prot-pr WAT-SC t70-white-adipose YP_665639.1 female
#> 534199:    <NA>  PROT    prot-pr WAT-SC t70-white-adipose YP_665639.1 female
#> 534200:    <NA>  PROT    prot-pr WAT-SC t70-white-adipose YP_665639.1 female
#>         comparison_group    p_value adj_p_value       logFC   logFC_se
#>      1:               1w 0.16522229           1  0.14138741 0.09855839
#>      2:               2w 0.02112377           1  0.24440461 0.09855839
#>      3:               4w 0.55497400           1  0.05906750 0.09855839
#>      4:               8w 0.71344616           1  0.03665581 0.09855839
#>      5:               1w 0.42333729           1 -0.06007221 0.07366201
#>     ---                                                               
#> 534196:               8w 0.04768560           1  0.34600052 0.16711151
#> 534197:               1w 0.37861543           1 -0.15957201 0.17832136
#> 534198:               2w 0.02287721           1 -0.42978938 0.17832136
#> 534199:               4w 0.39379681           1 -0.15448337 0.17832136
#> 534200:               8w 0.87194415           1 -0.02901243 0.17832136
#>             tscore covariates numNAs comparison_average_intensity
#>      1:  1.4345548         NA      2                   0.33459095
#>      2:  2.4797951         NA      2                   0.43760815
#>      3:  0.5993148         NA      2                   0.25227104
#>      4:  0.3719197         NA      2                   0.22985935
#>      5: -0.8155115         NA      2                   0.25880543
#>     ---                                                          
#> 534196:  2.0704769         NA      0                   0.18543879
#> 534197: -0.8948564         NA      0                  -0.22374709
#> 534198: -2.4101958         NA      0                  -0.49396446
#> 534199: -0.8663201         NA      0                  -0.21865845
#> 534200: -0.1626974         NA      0                  -0.09318751
#>         reference_average_intensity selection_fdr
#>      1:                  0.19320354     1.0000000
#>      2:                  0.19320354     1.0000000
#>      3:                  0.19320354     1.0000000
#>      4:                  0.19320354     1.0000000
#>      5:                  0.31887765     1.0000000
#>     ---                                          
#> 534196:                 -0.16056174     0.1785294
#> 534197:                 -0.06417508     0.1785294
#> 534198:                 -0.06417508     0.1785294
#> 534199:                 -0.06417508     0.1785294
#> 534200:                 -0.06417508     0.1785294

if (FALSE) {
# Return METHYL and ATAC differential analysis results for gastrocnemius 
combine_da_results(tissues="SKM-GN", 
                   assays=c("ATAC","METHYL"),
                   include_epigen=TRUE)
}