Pathway enrichment network

Plot an interactive network of pathway enrichments.

Usage

enrichment_network_vis(
  pw_enrich_res,
  feature_to_gene = MotrpacRatTraining6moData::FEATURE_TO_GENE,
  intersection_id_type = "ensembl_gene",
  similarity_cutoff = 0.375,
  adj_pval_cutoff = 0.1,
  title = NULL,
  label_nodes = TRUE,
  add_group_label_nodes = FALSE,
  parent_pathways = MotrpacRatTraining6moData::PATHWAY_PARENTS,
  multitissue_pathways_only = FALSE,
  include_metab_singletons = TRUE,
  similarity_scores_file = NULL,
  out_html = "/dev/null",
  overwrite_html = TRUE,
  return_html = FALSE,
  return_graph_for_cytoscape = FALSE,
  verbose = TRUE
)

Arguments

pw_enrich_res

data frame of enrichment results. May be a subset of MotrpacRatTraining6moData::GRAPH_PW_ENRICH or a data frame returned from cluster_pathway_enrichment(). Columns must include "adj_p_value", "ome", "tissue", "intersection", "computed_p_value", "term_name", "term_id".

feature_to_gene

data frame, map between intersection_id_type and gene symbols. Columns must include "gene_symbol" and "ensembl_gene" if intersection_id_type == "ensembl_gene". MotrpacRatTraining6moData::FEATURE_TO_GENE by default.

intersection_id_type

character, type of gene identifier used to define pw_enrich_res$intersection, either "ensembl_gene" or "gene_symbol"

similarity_cutoff

numeric, edges are drawn between pairs of pathways if they have a similarity metric above this value. 0.375 by default.

adj_pval_cutoff

numeric, pathway enrichments are only considered for the network if the corresponding adjusted p-value (pw_enrich_res$adj_p_value) is less than this value. 0.1 (10% FDR) by default.

title

character, plot title. NULL by default.

label_nodes

boolean, whether to label groups at all. TRUE by default. If label_nodes = FALSE, add_group_label_nodes is ignored.

add_group_label_nodes

boolean, whether to label groups with nodes. If FALSE, use a standard legend instead.

parent_pathways

named list, map of KEGG and REAC pathway term ID to parent pathway. Used to create labels for clusters of pathway enrichments. List names must correspond to values in pw_enrich_res$term_id. MotrpacRatTraining6moData::PATHWAY_PARENTS by default. If NULL, then only the pathway names are used to create cluster labels, which is quite meaningless is all pathway names are unique. In that case, it is recommended to set label_nodes to FALSE.

multitissue_pathways_only

boolean. If TRUE, only include pathways in the network if they are significantly enriched in more than one tissue in pw_enrich_res.

include_metab_singletons

boolean. If TRUE, include pathways enriched only by metabolites (ome METAB) as singleton nodes

similarity_scores_file

character or NULL, path in which to save pairwise pathway similarity scores as an RDS file. If this file exists, the pairwise similarity scores are loaded from the file instead of being recalculated. NULL by default.

out_html

character, output file for HTML. "/dev/null" by default.

overwrite_html

boolean, whether to overwrite out_html if it already exists. TRUE by default. If out_html exists and overwrite_html=FALSE, then the path out_html is returned.

return_html

boolean, whether to return the path to out_html. If FALSE, return the visNetwork object instead. FALSE by default.

return_graph_for_cytoscape

boolean. If TRUE, return an igraph graph object that can be exported to Cytoscape.

verbose

boolean, whether to print verbose output

Value

If return_graph_for_cytoscape=FALSE and return_html=FALSE, return a visNetwork graph. If return_graph_for_cytoscape=TRUE, return an igraph object. If return_html=TRUE, return the path to the HTML file in which the interactive visNetwork was saved.

See also

Examples

# Example 1: Plot an interactive network of pathway enrichments from the HEART:8w_F1_M1 node,
# i.e., features that are up-regulated in both males and females after 8 weeks of training
enrich_res = MotrpacRatTraining6moData::GRAPH_PW_ENRICH
enrich_res = enrich_res[enrich_res$cluster == "HEART:8w_F1_M1",]
enrichment_network_vis(enrich_res, add_group_label_nodes = TRUE) 
#> Formatting inputs...
#> Subsetting feature-to-gene map...
#> Calculating similarity metric between pairs of enriched pathways...
#> Constructing graph...
#> Elapsed time: 12.328s
#> 

if (FALSE) {
# Example 2: Export the above network to Cytoscape.
# Cytoscape must be running locally for this to work. 
library(RCy3)
g = enrichment_network_vis(enrich_res, return_graph_for_cytoscape = T) 
RCy3::cytoscapePing()
RCy3::createNetworkFromIgraph(g, new.title='HEART:8w_F1_M1')
}

# Example 3: Plot an interactive network of pathway enrichments corresponding to 
# features that are up-regulated in both sexes at 8 weeks in any of the 3 muscle tissues. 
# Only include pathways that are enriched in at least 2 muscle tissues.
# Plot the color legend on the left instead of using colored labels pointing to nodes 
enrich_res = MotrpacRatTraining6moData::GRAPH_PW_ENRICH
enrich_res = enrich_res[enrich_res$tissue %in% c("SKM-GN","SKM-VL","HEART") &
                         grepl(":8w_F1_M1",enrich_res$cluster),]
enrichment_network_vis(enrich_res,
                       similarity_cutoff = 0.3,
                       title = "Muscle 8w_F1_M1",
                       add_group_label_nodes = FALSE,
                       multitissue_pathways_only = TRUE) 
#> Formatting inputs...
#> Subsetting feature-to-gene map...
#> Calculating similarity metric between pairs of enriched pathways...
#> Constructing graph...
#> Elapsed time: 4.512s
#>