Find human flanks

Find flanking sequence in human ortholog of PTM

Usage

find_flanks(x, fasta, type = "uniprot")

Arguments

x

character, human accession for PTM, e.g., "Q96QG7_S548s"

fasta

Biostrings::XStringSet object returned from reading in a human protein FASTA file with Biostrings::readAAStringSet(). Names of the Biostrings::XStringSet object should be set to the human protein accession, e.g., "Q96QG7". Can be the result of load_uniprot_human_fasta(), which returns the version of the Uniprot human protein FASTA used in the manuscript.

type

character, type of accession

Value

character, pipe-separated string of human flanking sequences for the PTM

See also

Examples

fasta = load_uniprot_human_fasta(scratchdir="/tmp")
x = "Q96QG7_S548s"
find_flanks(x, fasta)
#> [1] "TEDGMQESP______"