Format gene symbols

Format a string of comma-separated gene symbols for use in enrichment_network_vis().

Usage

format_gene_symbols(x, return_N = TRUE, collapse = TRUE)

Arguments

x

character, comma-separated gene symbols

return_N

boolean, whether to prepend the concatenated gene symbols with the number of unique gene symbols. TRUE by default.

collapse

boolean, whether to collapse the gene symbols into a comma-separated string. If not, return a vector of unique gene symbols. TRUE by default.

Value

either a string of comma-separated gene symbols, a vector of unique gene symbols, or a named list with two values ("N" and "genes") depending on the values of return_N and collapse.

See also