Immunoassay timewise differential analysis

For each panel and tissue, perform pairwise contrasts between each group of trained animals and the sex-matched control group. Analysis is performed separately for males and females.

Usage

immuno_timewise_da()

Value

a data frame with one row per immunoassay feature per contrast (usually 8 rows per feature):

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

dataset

character, LUMINEX panel

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

covariates

character, comma-separated list of adjustment variables

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

Examples

res = immuno_timewise_da()
#> Group 2w in rat-mag27plex:GCSF:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:EOTAXIN:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:GMCSF:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL1A:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:MIP1A:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL4:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL2:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL6:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:EGF:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL13:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL12P70:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IFNY:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL5:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL17:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:IL18:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:MCP1:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:KC:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:VEGF:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:FRACTALKINE:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:LIX:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:MIP2:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:TNFA:COLON:male has less than 2 values. Removing remaining samples in this group.
#> Group 2w in rat-mag27plex:RANTES:COLON:male has less than 2 values. Removing remaining samples in this group.