Load metabolomics differential analysis

Load training or timewise metabolomics differential analysis results for a single tissue. For timewise results, load from MotrpacRatTraining6moData::METAB_DA. For training results, load from GCS. These are the redundant, non-meta-analyzed differential analysis that were used as input for the metabolomics meta-analysis and meta-regression but not otherwise directly used in the manuscript. The differential analysis results provided by MotrpacRatTraining6moData::METAB_DA_METAREG are the non-redundant, meta-regression analysis results used in the manuscript.

Usage

load_metab_da(tissue, type = "timewise", ...)

Arguments

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

type

character, type of differential analysis summary statistics to return. One of 'timewise', 'training'. Note that the adjusted p-value from type="training" is already include in type="timewise" as selection_fdr.

...

additional arguments passed to get_file_from_url()

Value

data frame. For detailed column descriptions, see MotrpacRatTraining6moData::METAB_DA and MotrpacRatTraining6moData::TRAINING_DA.

See also

Examples

# Get timewise differential analysis results for the liver
res = load_metab_da("LIVER")

# Get training differential analysis results for the liver
res = load_metab_da("LIVER", type="training")