Load sample-level data

Load raw counts or normalized sample-level data. Optionally filter by training-regulated features. For epigenetic data (ATAC and METHYL), if training_regulated_only = FALSE, sample-level data is downloaded from Google Cloud Storage.

Usage

load_sample_data(
  tissue,
  assay,
  normalized = TRUE,
  training_regulated_only = FALSE,
  exclude_outliers = FALSE,
  scratchdir = ".",
  nrows = Inf,
  warnings = TRUE
)

Arguments

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

normalized

bool, whether to return normalized data. If FALSE, return raw counts.

training_regulated_only

bool, whether to filter features down to those training-regulated at 5% FDR

exclude_outliers

bool, whether to remove sample outliers specified by MotrpacRatTraining6moData::OUTLIERS

scratchdir

character, local directory in which to download data from Google Cloud Storage. Current working directory by default. Only applies if assay is ATAC or METHYL.

nrows

integer, number of rows to return. Defaults to Inf. Useful to return a subset of a large data frame for tests.

warnings

bool, whether to print warnings to the console. TRUE by default.

Value

a data.frame where features are in rows and numeric columns correspond to sample identifiers (vial labels)

See also

Examples

# Load RNA-seq raw counts for liver
data = load_sample_data("LIVER", "TRNSCRPT", normalized = FALSE)
#> TRNSCRPT_LIVER_RAW_COUNTS

# Load normalized metabolomics data for gastrocnemius
data = load_sample_data("SKM-GN", "METAB")
#> METAB SKM-GN normalized data from METAB_NORM_DATA_FLAT

# Load normalized protein abundance data for heart
data = load_sample_data("HEART", "PROT")
#> PROT_HEART_NORM_DATA

if (FALSE) { # \dontrun{
# Load ATAC-seq raw counts for hippocampus, excluding outliers 
data = load_sample_data("HIPPOC", 
                        "ATAC", 
                        exclude_outliers = TRUE, 
                        normalized = FALSE, 
                        scratchdir = "/tmp")
} # }