Load training differential analysis results

Load training summary statistics from differential analysis (DA) that tests the effect of training on each feature within each sex. Results are downloaded from a public Google Cloud Storage bucket.

Usage

load_training_da(assay, tissue, metareg = TRUE, ...)

Arguments

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

metareg

bool, whether to return the meta-regression results for METAB. Return the upstream differential analysis results if FALSE. TRUE by default.

...

additional arguments passed to get_file_from_url()

Value

data frame, or data table with argument return_data_table = TRUE

See also

Examples

res = load_training_da("ACETYL", "HEART")
if (FALSE) res = load_training_da("ATAC", "SKM-GN")
res = load_training_da("IMMUNO", "ADRNL")
res = load_training_da("METAB", "SKM-GN", metareg = TRUE)
res = load_training_da("METAB", "SKM-GN", metareg = FALSE)
if (FALSE) res = load_training_da("METHYL", "SKM-GN")
res = load_training_da("PHOSPHO", "LIVER")
res = load_training_da("PROT", "KIDNEY")
res = load_training_da("TRNSCRPT", "BLOOD")
res = load_training_da("UBIQ", "HEART")