Metabolomics timewise meta-analysis

Perform meta-analysis for repeated measurements and return non-redundant results. If there were no repeated measurements for a feature, the input summary statistics are returned. These results were not used in the manuscript.

Usage

metab_meta_analysis(tissue, input = NULL)

Arguments

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

input

r data frame, custom input. To see the expected format, look at a table returned by load_metab_da() with type="timewise".

Value

a named list, where res$meta_analysis_res is a list with one metafor::rma.uni() object per meta-analyzed feature, and res$merged_res is a data frame non-redundant measurements for all input features with the following columns:

feature_ID

character, MoTrPAC feature identifier

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

dataset

character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV ("METAB")

sex

character, one of 'male' or 'female'

is_targeted

logical, is this a targeted platform?

site

character, Chemical Analysis Site (CAS) name

metabolite_refmet

character, RefMet name of metabolite

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

zscore

numeric, z-score (either original t-score or z-score from meta-analysis)

cv

double, feature coefficient of variation in the dataset

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

I2

double, heterogeneity coefficient. Higher values mean the measurements between sites/platforms are more discordant.

QEp

double, heterogeneity p-value. A smaller p-value means the measurements between sites/platforms are more discordant.

min_nominal_p

double, minimum nominal p-value for this sex and time point across potentially multiple measurements in the input

Examples

# Perform meta-analysis on gastrocnemius timewise summary statistics
res = metab_meta_analysis("SKM-GN")
#> Computed 1000 meta-analysis results.