Meta-regression for a metabolite

Worker function wrapped by metab_meta_regression().

Usage

metabolite_meta_regression(metabolite_refmet, timewise, training = NULL)

Arguments

metabolite_refmet

character, RefMet name of metabolite

timewise

data frame with all timewise summary statistics

training

optional data frame with all training summary statistics

Value

a named list with meta-regression results, metabolite metadata, and training summary statistics if training is not NULL.

Details

We try multiple models per analyte:

Model 1:Two random effects factors if platform and is_targeted are not redundant. Default optimization.

Model 2:Two random effects factors if platform and is_targeted are not redundant with alternative optimization.

Model 3:platform and is_targeted are redundant. Use a single RE factor with default optimization. Also, use this if QMp is NA, which is an indication of over-parameterization of the model.

Model 4:platform and is_targeted are redundant. Use a single RE factor with alternative optimization.

Model 5:If all previous analyses failed, use a simple fixed-effects approach.

See https://www.metafor-project.org/doku.php/tips:models_with_or_without_intercept for some explanations. See https://www.publichealth.columbia.edu/research/population-health-methods/meta-regression for general background, including discussion on FE models for replicated experiments (as in our case).

"Use of a fixed effect meta-analysis model assumes all studies are estimating the same (common) treatment effect.":

Riley et al. https://www-bmj-com.stanford.idm.oclc.org/content/342/bmj.d549

Hypothetically, if all studies had an infinite sample size, there would be no differences due to chance and the differences in study estimates would completely disappear.

Unlike the other assays, having greater sampling variance in one of the sexes is not a problem for the model and we therefore do not need to split the sexes.

Most longitudinal meta-analyses include random effect terms per study. For examples see https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0164898, https://metafor-project.org/doku.php/tips:two_stage_analysis.