Proteomics timewise differential analysis

Timewise differential analysis for the proteome, phosphoproteome, acetylome, and ubiquitylome. Use limma to perform pairwise contrasts between each group of trained animals and the sex-matched control group for a single tissue and proteomics assay. Analysis is performed separately for males and females.

Usage

proteomics_timewise_da(assay, tissue, exclude_outliers = TRUE)

Arguments

assay

character, abbreviation for proteomics assay to be analyzed as defined by MotrpacRatTraining6moData::ASSAY_ABBREV. One of the following: PROT, PHOSPHO, ACETYL, UBIQ

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

exclude_outliers

bool, whether to remove sample outliers specified in MotrpacRatTraining6moData::OUTLIERS. TRUE by default.

Value

a data frame with one row per proteomics feature per contrast (usually 8 rows per feature):

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

removed_samples

character, comma-separated list of outliers (vial labels) removed from differential analysis

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

numNAs

integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals

Examples

# Run timewise differential analysis for heart proteins
res = proteomics_timewise_da("PROT","HEART")
#> PROT_HEART_NORM_DATA