repfdr wrapper

A general wrapper for running repfdr on a matrix of z-scores.

Usage

repfdr_wrapper(zscores, min_prior_for_config = 0.001, df = 20)

Arguments

zscores

A numeric matrix. Rows are analytes, columns are conditions (e.g., male week 1). Entries are z-scores.

min_prior_for_config

A number. The minimal prior probability on the configuration priors.

df

Integer. Degrees of freedom for fitting the estimated density f(z)

Value

A named list with two objects:

repfdr_clusters

a binary matrix representation of the clusters

repfdr_cluster_posteriors

a matrix with the cluster posteriors

Details

This wrapper runs inference for the two groups model in each column of zscores, and then run repfdr's EM process. To extract the fuzzy clustering solution, we exclude configurations whose prior probability is lower than min_prior_for_config.

Examples

# Simulate data with a single cluster
zcolnames = c(
  paste("male",c("1w","2w","4w","8w"),sep="_"),
  paste("female",c("1w","2w","4w","8w"),sep="_")
)
zscores = matrix(rnorm(80000),ncol=8,dimnames = list(1:10000,zcolnames))
repfdr_results = repfdr_wrapper(zscores, df=5)
#> Warning: In column 5 female_1w the estimated fraction of nulls is 1.
#> Warning: In column 7 female_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 female_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> Converged after 1 EM iterations.
#> Warning: In column 5 female_1w the estimated fraction of nulls is 1.
#> Warning: In column 7 female_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 female_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> EM iteration 3
#> EM iteration 4
#> EM iteration 5
#> EM iteration 6
#> EM iteration 7
#> EM iteration 8
#> EM iteration 9
#> EM iteration 10
#> EM iteration 11
#> EM iteration 12
#> EM iteration 13
#> EM iteration 14
#> EM iteration 15
#> EM iteration 16
#> EM iteration 17
#> EM iteration 18
#> EM iteration 19
#> EM iteration 20
#> EM iteration 21
#> EM iteration 22
#> EM iteration 23
#> EM iteration 24
#> EM iteration 25
#> EM iteration 26
#> EM iteration 27
#> EM iteration 28
#> EM iteration 29
#> EM iteration 30
#> EM iteration 31
#> EM iteration 32
#> EM iteration 33
#> EM iteration 34
#> EM iteration 35
#> EM iteration 36
#> EM iteration 37
#> EM iteration 38
#> EM iteration 39
#> EM iteration 40
#> EM iteration 41
#> EM iteration 42
#> EM iteration 43
#> EM iteration 44
#> EM iteration 45
#> EM iteration 46
#> EM iteration 47
#> EM iteration 48
#> EM iteration 49
#> EM iteration 50
#> EM iteration 51
#> EM iteration 52
#> EM iteration 53
#> EM iteration 54
#> EM iteration 55
#> EM iteration 56
#> EM iteration 57
#> EM iteration 58
#> EM iteration 59
#> EM iteration 60
#> EM iteration 61
#> EM iteration 62
#> EM iteration 63
#> EM iteration 64
#> EM iteration 65
#> EM iteration 66
#> EM iteration 67
#> EM iteration 68
#> EM iteration 69
#> EM iteration 70
#> EM iteration 71
#> EM iteration 72
#> EM iteration 73
#> EM iteration 74
#> EM iteration 75
#> EM iteration 76
#> EM iteration 77
#> EM iteration 78
#> EM iteration 79
#> EM iteration 80
#> EM iteration 81
#> EM iteration 82
#> EM iteration 83
#> EM iteration 84
#> EM iteration 85
#> EM iteration 86
#> EM iteration 87
#> EM iteration 88
#> EM iteration 89
#> EM iteration 90
#> EM iteration 91
#> EM iteration 92
#> EM iteration 93
#> EM iteration 94
#> EM iteration 95
#> EM iteration 96
#> EM iteration 97
#> EM iteration 98
#> EM iteration 99
#> EM iteration 100
#> EM iteration 101
#> EM iteration 102
#> EM iteration 103
#> EM iteration 104
#> EM iteration 105
#> EM iteration 106
#> EM iteration 107
#> EM iteration 108
#> EM iteration 109
#> EM iteration 110
#> EM iteration 111
#> EM iteration 112
#> EM iteration 113
#> EM iteration 114
#> EM iteration 115
#> EM iteration 116
#> EM iteration 117
#> EM iteration 118
#> EM iteration 119
#> EM iteration 120
#> EM iteration 121
#> EM iteration 122
#> EM iteration 123
#> EM iteration 124
#> EM iteration 125
#> EM iteration 126
#> EM iteration 127
#> EM iteration 128
#> EM iteration 129
#> EM iteration 130
#> EM iteration 131
#> EM iteration 132
#> EM iteration 133
#> EM iteration 134
#> EM iteration 135
#> EM iteration 136
#> EM iteration 137
#> EM iteration 138
#> EM iteration 139
#> EM iteration 140
#> EM iteration 141
#> EM iteration 142
#> EM iteration 143
#> EM iteration 144
#> EM iteration 145
#> EM iteration 146
#> EM iteration 147
#> EM iteration 148
#> EM iteration 149
#> EM iteration 150
#> EM iteration 151
#> EM iteration 152
#> EM iteration 153
#> EM iteration 154
#> EM iteration 155
#> EM iteration 156
#> EM iteration 157
#> EM iteration 158
#> EM iteration 159
#> EM iteration 160
#> EM iteration 161
#> EM iteration 162
#> EM iteration 163
#> EM iteration 164
#> EM iteration 165
#> EM iteration 166
#> EM iteration 167
#> EM iteration 168
#> EM iteration 169
#> EM iteration 170
#> EM iteration 171
#> EM iteration 172
#> EM iteration 173
#> EM iteration 174
#> EM iteration 175
#> EM iteration 176
#> EM iteration 177
#> EM iteration 178
#> EM iteration 179
#> EM iteration 180
#> EM iteration 181
#> EM iteration 182
#> EM iteration 183
#> EM iteration 184
#> EM iteration 185
#> EM iteration 186
#> EM iteration 187
#> EM iteration 188
#> EM iteration 189
#> EM iteration 190
#> EM iteration 191
#> EM iteration 192
#> EM iteration 193
#> EM iteration 194
#> EM iteration 195
#> EM iteration 196
#> EM iteration 197
#> EM iteration 198
#> EM iteration 199
#> EM iteration 200
#> EM iteration 201
#> EM iteration 202
#> EM iteration 203
#> EM iteration 204
#> EM iteration 205
#> EM iteration 206
#> EM iteration 207
#> EM iteration 208
#> EM iteration 209
#> EM iteration 210
#> EM iteration 211
#> EM iteration 212
#> EM iteration 213
#> EM iteration 214
#> EM iteration 215
#> EM iteration 216
#> EM iteration 217
#> EM iteration 218
#> EM iteration 219
#> EM iteration 220
#> EM iteration 221
#> EM iteration 222
#> EM iteration 223
#> EM iteration 224
#> EM iteration 225
#> EM iteration 226
#> EM iteration 227
#> EM iteration 228
#> EM iteration 229
#> EM iteration 230
#> EM iteration 231
#> EM iteration 232
#> EM iteration 233
#> EM iteration 234
#> EM iteration 235
#> EM iteration 236
#> EM iteration 237
#> EM iteration 238
#> EM iteration 239
#> EM iteration 240
#> EM iteration 241
#> EM iteration 242
#> EM iteration 243
#> EM iteration 244
#> EM iteration 245
#> EM iteration 246
#> EM iteration 247
#> EM iteration 248
#> EM iteration 249
#> EM iteration 250
#> EM iteration 251
#> EM iteration 252
#> EM iteration 253
#> EM iteration 254
#> EM iteration 255
#> EM iteration 256
#> EM iteration 257
#> EM iteration 258
#> EM iteration 259
#> EM iteration 260
#> EM iteration 261
#> EM iteration 262
#> EM iteration 263
#> EM iteration 264
#> EM iteration 265
#> EM iteration 266
#> EM iteration 267
#> EM iteration 268
#> EM iteration 269
#> EM iteration 270
#> EM iteration 271
#> EM iteration 272
#> EM iteration 273
#> EM iteration 274
#> EM iteration 275
#> EM iteration 276
#> EM iteration 277
#> EM iteration 278
#> EM iteration 279
#> EM iteration 280
#> EM iteration 281
#> EM iteration 282
#> EM iteration 283
#> EM iteration 284
#> EM iteration 285
#> EM iteration 286
#> EM iteration 287
#> EM iteration 288
#> EM iteration 289
#> EM iteration 290
#> EM iteration 291
#> EM iteration 292
#> EM iteration 293
#> EM iteration 294
#> EM iteration 295
#> EM iteration 296
#> EM iteration 297
#> EM iteration 298
#> EM iteration 299
#> EM iteration 300
#> EM iteration 301
#> EM iteration 302
#> EM iteration 303
#> EM iteration 304
#> EM iteration 305
#> EM iteration 306
#> EM iteration 307
#> EM iteration 308
#> EM iteration 309
#> EM iteration 310
#> EM iteration 311
#> EM iteration 312
#> EM iteration 313
#> EM iteration 314
#> EM iteration 315
#> EM iteration 316
#> EM iteration 317
#> EM iteration 318
#> EM iteration 319
#> EM iteration 320
#> EM iteration 321
#> EM iteration 322
#> EM iteration 323
#> EM iteration 324
#> EM iteration 325
#> EM iteration 326
#> EM iteration 327
#> EM iteration 328
#> EM iteration 329
#> EM iteration 330
#> EM iteration 331
#> EM iteration 332
#> EM iteration 333
#> EM iteration 334
#> EM iteration 335
#> EM iteration 336
#> EM iteration 337
#> EM iteration 338
#> EM iteration 339
#> EM iteration 340
#> EM iteration 341
#> EM iteration 342
#> EM iteration 343
#> EM iteration 344
#> EM iteration 345
#> EM iteration 346
#> EM iteration 347
#> EM iteration 348
#> EM iteration 349
#> EM iteration 350
#> EM iteration 351
#> EM iteration 352
#> EM iteration 353
#> EM iteration 354
#> EM iteration 355
#> EM iteration 356
#> EM iteration 357
#> EM iteration 358
#> EM iteration 359
#> EM iteration 360
#> EM iteration 361
#> EM iteration 362
#> EM iteration 363
#> EM iteration 364
#> EM iteration 365
#> EM iteration 366
#> EM iteration 367
#> EM iteration 368
#> EM iteration 369
#> EM iteration 370
#> EM iteration 371
#> EM iteration 372
#> EM iteration 373
#> EM iteration 374
#> EM iteration 375
#> EM iteration 376
#> EM iteration 377
#> EM iteration 378
#> EM iteration 379
#> EM iteration 380
#> EM iteration 381
#> EM iteration 382
#> EM iteration 383
#> EM iteration 384
#> EM iteration 385
#> EM iteration 386
#> EM iteration 387
#> EM iteration 388
#> EM iteration 389
#> EM iteration 390
#> EM iteration 391
#> EM iteration 392
#> EM iteration 393
#> EM iteration 394
#> EM iteration 395
#> EM iteration 396
#> EM iteration 397
#> EM iteration 398
#> EM iteration 399
#> EM iteration 400
#> EM iteration 401
#> EM iteration 402
#> EM iteration 403
#> EM iteration 404
#> EM iteration 405
#> EM iteration 406
#> EM iteration 407
#> EM iteration 408
#> EM iteration 409
#> EM iteration 410
#> EM iteration 411
#> EM iteration 412
#> EM iteration 413
#> EM iteration 414
#> EM iteration 415
#> EM iteration 416
#> EM iteration 417
#> EM iteration 418
#> EM iteration 419
#> EM iteration 420
#> EM iteration 421
#> EM iteration 422
#> EM iteration 423
#> EM iteration 424
#> EM iteration 425
#> EM iteration 426
#> EM iteration 427
#> EM iteration 428
#> EM iteration 429
#> EM iteration 430
#> EM iteration 431
#> EM iteration 432
#> EM iteration 433
#> EM iteration 434
#> EM iteration 435
#> EM iteration 436
#> EM iteration 437
#> EM iteration 438
#> EM iteration 439
#> EM iteration 440
#> EM iteration 441
#> EM iteration 442
#> EM iteration 443
#> EM iteration 444
#> EM iteration 445
#> EM iteration 446
#> EM iteration 447
#> EM iteration 448
#> EM iteration 449
#> EM iteration 450
#> EM iteration 451
#> EM iteration 452
#> EM iteration 453
#> EM iteration 454
#> EM iteration 455
#> EM iteration 456
#> EM iteration 457
#> EM iteration 458
#> EM iteration 459
#> EM iteration 460
#> EM iteration 461
#> EM iteration 462
#> EM iteration 463
#> EM iteration 464
#> EM iteration 465
#> EM iteration 466
#> EM iteration 467
#> EM iteration 468
#> EM iteration 469
#> EM iteration 470
#> EM iteration 471
#> EM iteration 472
#> EM iteration 473
#> EM iteration 474
#> EM iteration 475
#> EM iteration 476
#> EM iteration 477
#> EM iteration 478
#> EM iteration 479
#> EM iteration 480
#> EM iteration 481
#> EM iteration 482
#> EM iteration 483
#> EM iteration 484
#> EM iteration 485
#> EM iteration 486
#> EM iteration 487
#> EM iteration 488
#> EM iteration 489
#> EM iteration 490
#> EM iteration 491
#> EM iteration 492
#> EM iteration 493
#> EM iteration 494
#> EM iteration 495
#> EM iteration 496
#> EM iteration 497
#> EM iteration 498
#> EM iteration 499
#> EM iteration 500
#> Warning: EM did not converge within tolerance after maximum number of iterations specified.
# in this example all configurations are null, 
# thus the  posteriors of the null cluster (all zeroes) are very high:
quantile(repfdr_results$repfdr_cluster_posteriors[,"00000000"])
#>         0%        25%        50%        75%       100% 
#> 0.04542362 0.95560012 0.98366441 0.99696815 1.00000000 
# now add a cluster with a strong signal and rerun
zscores[1:500,1:4] = zscores[1:500,1:4] + 5
# When the data are "clean" (e.g., a mixture of two gaussians), we do not
# need a high df in the two-groups model estimation
# (default is 20, consider at least 10 when analyzing real data)
repfdr_results = repfdr_wrapper(zscores, df=5)
#> Warning: In column  1   male_1w ,f(z) misfit =  2.7 . Rerun with increased df
#> Warning: In column  4   male_8w ,f(z) misfit =  1.9 . Rerun with increased df
#> Warning: In column 5 female_1w the estimated fraction of nulls is 1.
#> Warning: In column 7 female_4w the estimated fraction of nulls is 1.
#> Warning: In column 8 female_8w the estimated fraction of nulls is 1.
#> EM iteration 2
#> EM iteration 3
#> EM iteration 4
#> EM iteration 5
#> EM iteration 6
#> EM iteration 7
#> EM iteration 8
#> EM iteration 9
#> EM iteration 10
#> EM iteration 11
#> EM iteration 12
#> EM iteration 13
#> EM iteration 14
#> EM iteration 15
#> EM iteration 16
#> EM iteration 17
#> EM iteration 18
#> EM iteration 19
#> EM iteration 20
#> EM iteration 21
#> EM iteration 22
#> EM iteration 23
#> EM iteration 24
#> EM iteration 25
#> EM iteration 26
#> EM iteration 27
#> EM iteration 28
#> EM iteration 29
#> EM iteration 30
#> EM iteration 31
#> EM iteration 32
#> EM iteration 33
#> EM iteration 34
#> EM iteration 35
#> EM iteration 36
#> EM iteration 37
#> EM iteration 38
#> EM iteration 39
#> EM iteration 40
#> EM iteration 41
#> EM iteration 42
#> EM iteration 43
#> EM iteration 44
#> EM iteration 45
#> EM iteration 46
#> EM iteration 47
#> EM iteration 48
#> EM iteration 49
#> EM iteration 50
#> EM iteration 51
#> EM iteration 52
#> EM iteration 53
#> EM iteration 54
#> EM iteration 55
#> EM iteration 56
#> EM iteration 57
#> EM iteration 58
#> EM iteration 59
#> EM iteration 60
#> EM iteration 61
#> EM iteration 62
#> EM iteration 63
#> EM iteration 64
#> EM iteration 65
#> EM iteration 66
#> EM iteration 67
#> EM iteration 68
#> EM iteration 69
#> EM iteration 70
#> EM iteration 71
#> EM iteration 72
#> EM iteration 73
#> EM iteration 74
#> EM iteration 75
#> EM iteration 76
#> EM iteration 77
#> EM iteration 78
#> EM iteration 79
#> EM iteration 80
#> EM iteration 81
#> EM iteration 82
#> EM iteration 83
#> EM iteration 84
#> EM iteration 85
#> EM iteration 86
#> EM iteration 87
#> EM iteration 88
#> EM iteration 89
#> EM iteration 90
#> EM iteration 91
#> EM iteration 92
#> EM iteration 93
#> EM iteration 94
#> EM iteration 95
#> EM iteration 96
#> EM iteration 97
#> EM iteration 98
#> EM iteration 99
#> EM iteration 100
#> EM iteration 101
#> EM iteration 102
#> EM iteration 103
#> EM iteration 104
#> EM iteration 105
#> EM iteration 106
#> EM iteration 107
#> EM iteration 108
#> EM iteration 109
#> EM iteration 110
#> EM iteration 111
#> EM iteration 112
#> EM iteration 113
#> EM iteration 114
#> EM iteration 115
#> EM iteration 116
#> EM iteration 117
#> EM iteration 118
#> EM iteration 119
#> EM iteration 120
#> EM iteration 121
#> EM iteration 122
#> EM iteration 123
#> EM iteration 124
#> EM iteration 125
#> EM iteration 126
#> EM iteration 127
#> EM iteration 128
#> EM iteration 129
#> EM iteration 130
#> EM iteration 131
#> EM iteration 132
#> EM iteration 133
#> EM iteration 134
#> EM iteration 135
#> EM iteration 136
#> EM iteration 137
#> EM iteration 138
#> EM iteration 139
#> EM iteration 140
#> EM iteration 141
#> EM iteration 142
#> EM iteration 143
#> EM iteration 144
#> EM iteration 145
#> EM iteration 146
#> EM iteration 147
#> EM iteration 148
#> EM iteration 149
#> EM iteration 150
#> EM iteration 151
#> EM iteration 152
#> EM iteration 153
#> EM iteration 154
#> EM iteration 155
#> EM iteration 156
#> EM iteration 157
#> EM iteration 158
#> EM iteration 159
#> EM iteration 160
#> EM iteration 161
#> EM iteration 162
#> EM iteration 163
#> EM iteration 164
#> EM iteration 165
#> EM iteration 166
#> EM iteration 167
#> EM iteration 168
#> EM iteration 169
#> EM iteration 170
#> EM iteration 171
#> EM iteration 172
#> EM iteration 173
#> EM iteration 174
#> EM iteration 175
#> EM iteration 176
#> EM iteration 177
#> EM iteration 178
#> EM iteration 179
#> EM iteration 180
#> EM iteration 181
#> EM iteration 182
#> EM iteration 183
#> EM iteration 184
#> EM iteration 185
#> EM iteration 186
#> EM iteration 187
#> EM iteration 188
#> EM iteration 189
#> EM iteration 190
#> EM iteration 191
#> EM iteration 192
#> EM iteration 193
#> EM iteration 194
#> EM iteration 195
#> EM iteration 196
#> EM iteration 197
#> EM iteration 198
#> EM iteration 199
#> EM iteration 200
#> EM iteration 201
#> EM iteration 202
#> EM iteration 203
#> EM iteration 204
#> EM iteration 205
#> EM iteration 206
#> EM iteration 207
#> EM iteration 208
#> EM iteration 209
#> EM iteration 210
#> EM iteration 211
#> EM iteration 212
#> EM iteration 213
#> EM iteration 214
#> EM iteration 215
#> EM iteration 216
#> EM iteration 217
#> EM iteration 218
#> EM iteration 219
#> EM iteration 220
#> EM iteration 221
#> EM iteration 222
#> EM iteration 223
#> EM iteration 224
#> EM iteration 225
#> EM iteration 226
#> EM iteration 227
#> EM iteration 228
#> EM iteration 229
#> EM iteration 230
#> EM iteration 231
#> EM iteration 232
#> EM iteration 233
#> EM iteration 234
#> EM iteration 235
#> EM iteration 236
#> EM iteration 237
#> EM iteration 238
#> EM iteration 239
#> EM iteration 240
#> EM iteration 241
#> EM iteration 242
#> EM iteration 243
#> EM iteration 244
#> EM iteration 245
#> EM iteration 246
#> Converged after 245 EM iterations.
# look at the null cluster after adding the signal above
quantile(repfdr_results$repfdr_cluster_posteriors[,"00000000"],probs=c(0.05,0.1,0.5))
#>         5%        10%        50% 
#> 0.07926923 0.92329821 0.99725019 
# now the posteriors of the first 500 rows, 
# with respect to the "planted" cluster should have high posteriors:
quantile(repfdr_results$repfdr_cluster_posteriors[1:500,"11110000"])
#>        0%       25%       50%       75%      100% 
#> 0.0000000 0.9830629 0.9991024 1.0000000 1.0000000