Run single iteration of pathway enrichment

Run single iteration of pathway enrichment (FELLA for METAB or gprofiler for all other omes) NOT intended to be run directly. Run within cluster_pathway_enrichment().

Usage

run_single_enrichment(
  feature_to_gene_map,
  universe,
  cluster_res,
  gene_identifier_type,
  databases,
  iterations,
  i,
  min_input_set_size,
  fella.data,
  fella_method,
  logfile = "/dev/null",
  maxattempt = 50
)

Arguments

feature_to_gene_map

data table, map between intersection_id_type and gene symbols. Columns must include "feature_ID", "gene_symbol", "ensembl_gene", and "kegg_id".

universe

character vector of genes in the universe or background

cluster_res

data frame output by check_cluster_res_format()

gene_identifier_type

column in feature_to_gene_map that matches the gene identifier type in universe

databases

character vector of g:Profiler pathway databases to query. "KEGG" and "REAC" (REACTOME) by default. Current options include: GO (GO:BP, GO:MF, GO:CC to select a particular GO branch), KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP. See gprofiler2 documentation for an up-to-date list.

iterations

data frame passed from cluster_pathway_enrichment() that defines each iteration of this function

i

integer index passed from cluster_pathway_enrichment() that specifies which row of iterations to use

min_input_set_size

integer, input must have this minimum number of unique mappable gene IDs to attempt enrichment with gprofiler2::gost()

fella.data

FELLA database returned by FELLA::loadKEGGdata() passed to run_fella()

fella_method

character, enrichment method for FELLA::enrich(), one of "hypergeom" or "diffusion", passed to run_fella()

logfile

optional character, path to log of failed iterations

maxattempt

integer, max number of consecutive null results from gprofiler2::gost() before giving up

Value

data table with enrichment results, or NULL if no enrichment results were returned

See also