Combine differential analysis results

Combine differential analysis results from the specified tissues and omes(s)/assay(s). If no tissues or omes are specified, all differential analysis results are returned.

Usage

combine_da_results(
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  metareg = TRUE,
  include_epigen = FALSE,
  scratchdir = "."
)

Arguments

tissues

optional character vector of tissue abbreviations, one of MotrpacRatTraining6moData::TISSUE_ABBREV.

assays

optional character vector of assay abbreviations, one of MotrpacRatTraining6moData::ASSAY_ABBREV

metareg

bool, whether to use the meta-regression results for METAB. Use the upstream differential analysis results if FALSE. TRUE by default.

include_epigen

bool, whether to include the full ATAC or METHYL differential analysis results from Google Cloud Storage. Only relevant if assays includes "ATAC" or "METHYL". FALSE by default.

scratchdir

character, local directory in which to download data from the web. Current working directory by default. Only relevant if assays includes "ATAC" or "METHYL".

Value

data frame. Depending on the specified assays, some of these columns may not be included:

feature

character, unique feature identifier in the format 'MotrpacRatTraining6moData::ASSAY_ABBREV;MotrpacRatTraining6moData::TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See MotrpacRatTraining6moData::REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of MotrpacRatTraining6moData::ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of MotrpacRatTraining6moData::TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

covariates

character, comma-separated list of adjustment variables or NA

numNAs

integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

dataset

character, immune panel, metabolomics platform, or ATAC-seq dataset name

site

character, Chemical Analysis Site (CAS) name. METAB only

is_targeted

logical, is this a targeted platform? METAB only

metabolite_refmet

character, RefMet name of metabolite. METAB only

cv

double, feature coefficient of variation in the dataset. METAB only

metabolite

character, name of metabolite as appears in the CAS's data. METAB only

control_cv

double, feature coefficient of variation in the dataset. METAB only

mz

double, mass over charge. METAB only

rt

double, retention time. METAB only

neutral_mass

double, neutral mass. METAB only

meta_reg_het_p

double, for metabolites with multiple measurements, the meta-regression heterogeneity p-value, where a smaller p-value indicates more disagreement between platforms. One value per feature (not per training group). METAB only

meta_reg_pvalue

double, for metabolites with multiple measurements, the meta-regression p-value. One value per feature (not per training group). METAB only

shrunk_logFC

double, log fold-change with shrinkage applied

shrunk_logFC_se

double, standard error of the shrunken log fold-change

zscore

double, z statistic

removed_samples

character, comma-separated list of vial labels excluded from differential analysis or NA

comparison_average_intensity_se

double, standard error of 'comparison_average_intensity'

reference_average_intensity_se

double, standard error of 'comparison_average_intensity'

Chr

integer, chromosome. METHYL only

Locus

character, base pair range of feature. METHYL only

EntrezID

character, Entrez ID of closest gene. METHYL only

Symbol

character, gene symbol of closest gene. METHYL only

Examples

if (FALSE) { # \dontrun{
# Return all non-epigenetic differential analysis results, 
# including meta-regression results for metabolomics
data = combine_da_results()

# Return all global proteomics differential analysis results
data = combine_da_results(assays="PROT")

# Return METHYL and ATAC differential analysis results for gastrocnemius 
data = combine_da_results(tissues="SKM-GN", 
                          assays=c("ATAC","METHYL"),
                          include_epigen=TRUE)
} # }