Combine normalized sample-level data

Combine normalized sample-level data from the specified tissues and omes(s)/assay(s). If no tissues or omes are specified, all data is returned. In order to combine data from different tissues and data types, sample-specific vial labels are converted to animal-specific Participant IDs (PIDs).

Usage

combine_normalized_data(
  tissues = MotrpacRatTraining6moData::TISSUE_ABBREV,
  assays = MotrpacRatTraining6moData::ASSAY_ABBREV,
  include_epigen = FALSE,
  scratchdir = ".",
  training_regulated_only = FALSE,
  exclude_outliers = FALSE,
  nrows = Inf
)

Arguments

tissues

optional character vector of tissue abbreviations, one of MotrpacRatTraining6moData::TISSUE_ABBREV.

assays

optional character vector of assay abbreviations, one of MotrpacRatTraining6moData::ASSAY_ABBREV

include_epigen

bool, whether to include the full ATAC or METHYL differential analysis results from Google Cloud Storage. Only relevant if assays includes "ATAC" or "METHYL". FALSE by default.

scratchdir

character, local directory in which to download data from the web. Current working directory by default. Only relevant if assays includes "ATAC" or "METHYL".

training_regulated_only

bool, whether to filter features down to those training-regulated at 5% FDR

exclude_outliers

bool, whether to remove sample outliers specified by MotrpacRatTraining6moData::OUTLIERS

nrows

integer, number of rows to return from each dataset. Defaults to Inf. Useful to return a subset of a large data frame for tests.

Value

data frame with features in rows and Participant IDs (PIDs) in columns

See also

Examples

if (FALSE) {
# Return all normalized RNA-seq data
data = combine_normalized_data(assays = "TRNSCRPT")
}

# Return all normalized proteomics data. Exclude outliers 
data = combine_normalized_data(assays = c("PROT","UBIQ","PHOSPHO","ACETYL"),
                               exclude_outliers = TRUE)
#> PROT_CORTEX_NORM_DATA
#> PROT_HEART_NORM_DATA
#> PROT_KIDNEY_NORM_DATA
#> PROT_LIVER_NORM_DATA
#> PROT_LUNG_NORM_DATA
#> PROT_SKMGN_NORM_DATA
#> PROT_WATSC_NORM_DATA
#> UBIQ_HEART_NORM_DATA
#> UBIQ_LIVER_NORM_DATA
#> PHOSPHO_CORTEX_NORM_DATA
#> PHOSPHO_HEART_NORM_DATA
#> PHOSPHO_KIDNEY_NORM_DATA
#> PHOSPHO_LIVER_NORM_DATA
#> PHOSPHO_LUNG_NORM_DATA
#> PHOSPHO_SKMGN_NORM_DATA
#> PHOSPHO_WATSC_NORM_DATA
#> ACETYL_HEART_NORM_DATA
#> ACETYL_LIVER_NORM_DATA

if (FALSE) {
# Return normalized ATAC-seq data for training-regulated features 
data = combine_normalized_data(assays = "ATAC", training_regulated_only = TRUE)

# Return normalized ATAC-seq data for the first 1000 features in each tissue 
data = combine_normalized_data(assays = "ATAC", 
                               nrows = 1000, 
                               scratchdir = "/tmp", 
                               include_epigen = TRUE)
}

# Return all normalized metabolomics data 
data = combine_normalized_data(assays = "METAB")
#> METAB ADRNL normalized data from METAB_NORM_DATA_FLAT
#> METAB BAT normalized data from METAB_NORM_DATA_FLAT
#> METAB BLOOD normalized data from METAB_NORM_DATA_FLAT
#> METAB COLON normalized data from METAB_NORM_DATA_FLAT
#> METAB CORTEX normalized data from METAB_NORM_DATA_FLAT
#> METAB HEART normalized data from METAB_NORM_DATA_FLAT
#> METAB HIPPOC normalized data from METAB_NORM_DATA_FLAT
#> METAB HYPOTH normalized data from METAB_NORM_DATA_FLAT
#> METAB KIDNEY normalized data from METAB_NORM_DATA_FLAT
#> METAB LIVER normalized data from METAB_NORM_DATA_FLAT
#> METAB LUNG normalized data from METAB_NORM_DATA_FLAT
#> METAB OVARY normalized data from METAB_NORM_DATA_FLAT
#> METAB PLASMA normalized data from METAB_NORM_DATA_FLAT
#> METAB SKM-GN normalized data from METAB_NORM_DATA_FLAT
#> METAB SKM-VL normalized data from METAB_NORM_DATA_FLAT
#> METAB SMLINT normalized data from METAB_NORM_DATA_FLAT
#> METAB SPLEEN normalized data from METAB_NORM_DATA_FLAT
#> METAB TESTES normalized data from METAB_NORM_DATA_FLAT
#> METAB VENACV normalized data from METAB_NORM_DATA_FLAT
#> METAB WAT-SC normalized data from METAB_NORM_DATA_FLAT