Make data frame numeric only

Set row names and remove all non-numeric columns. This is useful for reformatting data objects in MotrpacRatTraining6moData, e.g., MotrpacRatTraining6moData::PROT_HEART_NORM_DATA

Usage

df_to_numeric(df, rownames = "feature_ID")

Arguments

df

data frame with at least one numeric column

rownames

either NULL, a character specifying a column name in df, or a vector. Specifies target row names. Default: "feature_ID"

Value

numeric data frame

Examples

df = MotrpacRatTraining6moData::PROT_HEART_NORM_DATA
head(df)
#>                     feature     feature_ID tissue assay 90237015805 90245015805
#> 1                      <NA> XP_017456475.1  HEART  PROT     0.00769    -0.05952
#> 2                      <NA> XP_017447817.1  HEART  PROT     0.09326    -0.08879
#> 3                      <NA>    NP_446341.1  HEART  PROT    -0.11469     0.10623
#> 4                      <NA>    NP_071796.2  HEART  PROT     0.05903    -0.06490
#> 5                      <NA> NP_001157776.1  HEART  PROT     1.24616     0.63626
#> 6 PROT;HEART;XP_003754359.1 XP_003754359.1  HEART  PROT    -0.04140    -0.00964
#>   90441015805 90420015805 90289015805 90267015805 90218015805 90585015805
#> 1     0.10992    -0.00229     0.04674     0.06119     0.11000    -0.02633
#> 2     0.31763     0.07739    -0.06031     0.10648     0.14415     0.12522
#> 3     0.08138     0.08464    -0.16039    -0.09960    -0.09566    -0.02896
#> 4     0.05673    -0.07494     0.06263    -0.02216     0.10700    -0.07468
#> 5     0.15487    -0.39231    -0.64953     0.63982     0.12769    -0.46366
#> 6    -0.08768     0.10905    -0.01300     0.06778     0.05061     0.03325
#>   90432015802 90560015805 90430015805 90412015805 90217015805 90266015805
#> 1     0.08646    -0.06782    -0.00225     0.01435     0.07922    -0.01652
#> 2     0.13868    -0.03150     0.09258     0.08732     0.12027    -0.02472
#> 3    -0.11469     0.26114     0.14095    -0.03176    -0.16405    -0.13004
#> 4     0.05919    -0.09779     0.08112    -0.06524     0.09477    -0.03734
#> 5    -1.49219    -1.48255     0.14749    -0.36679    -0.34689    -0.00593
#> 6    -0.14632     0.00766    -0.04890     0.00656    -0.09305    -0.08272
#>   90223015805 90578015805 90449015805 90258015805 90283015805 90571015805
#> 1     0.11231    -0.02321     0.09507     0.00787     0.07536    -0.07617
#> 2     0.16095     0.07759     0.05649     0.19117    -0.02809     0.08864
#> 3    -0.07527    -0.01265    -0.14445     0.06341    -0.01666     0.28993
#> 4     0.09170    -0.10477     0.08221    -0.05721     0.05121    -0.12634
#> 5     0.29859    -0.57429    -0.30849    -0.10142    -0.03489    -0.38285
#> 6     0.03596    -0.03905    -0.00406     0.10679     0.04905     0.03972
#>   90281015805 90587015805 90222015805 90256015802 90239015805 90559015805
#> 1     0.07170     0.02030     0.09243    -0.03238     0.04958    -0.05368
#> 2     0.18576     0.09423     0.12652     0.04303     0.06640     0.08883
#> 3    -0.12354    -0.00561    -0.14562     0.05136    -0.01653     0.09173
#> 4     0.06304    -0.01797     0.05310    -0.06212     0.03588    -0.06982
#> 5     0.10696    -0.20935    -0.09217    -0.31013    -0.12688    -0.10168
#> 6     0.03327     0.02120    -0.00913    -0.06928    -0.03088    -0.01289
#>   90439015805 90406015805 90423015805 90255015802 90227015805 90567015805
#> 1     0.03511    -0.00482     0.07782     0.00998     0.12675    -0.10506
#> 2     0.13454    -0.01024     0.14147    -0.04832     0.26463    -0.00900
#> 3     0.04071    -0.08929    -0.03453     0.15072    -0.15487     0.16851
#> 4     0.00694    -0.02401     0.07256    -0.04749     0.11555    -0.08556
#> 5    -0.06460    -0.53958     0.02972    -0.36775     0.32162    -0.01294
#> 6    -0.03503     0.00117    -0.03446     0.10633     0.07847    -0.03708
#>   90229015805 90576015805 90280015805 90252015805 90445015802 90254015805
#> 1     0.07816    -0.04595     0.10060    -0.04869     0.07865     0.01499
#> 2     0.03422    -0.05347     0.17520     0.05656     0.24797     0.04646
#> 3    -0.01903    -0.01581    -0.22340    -0.06180    -0.09393     0.21671
#> 4     0.04804    -0.07206     0.08251    -0.06693     0.04238    -0.05073
#> 5    -0.06972    -0.66205     0.14655    -0.41869     0.03080    -0.41675
#> 6    -0.07135    -0.00649    -0.00404    -0.05693    -0.03708     0.08520
#>   90421015805 90416015805 90426015805 90265015805 90292015805 90574015802
#> 1     0.06699    -0.00042     0.07749    -0.02423     0.11087    -0.06783
#> 2     0.15554    -0.09590     0.16798     0.07771     0.13045    -0.06292
#> 3     0.05044     0.05260     0.05966    -0.06085    -0.17454    -0.00723
#> 4     0.05478    -0.05785     0.08925    -0.06092     0.07279    -0.07737
#> 5    -0.33470    -0.25958     0.14447    -0.23248    -0.04627    -0.57912
#> 6    -0.03783     0.05743    -0.05527    -0.06289    -0.04862    -0.02176
#>   90234015802 90410015805 90444015805 90569015802 90232015805 90251015805
#> 1     0.03175     0.03331     0.05923    -0.01780     0.05311     0.01011
#> 2     0.15962     0.09036     0.19256    -0.03660     0.00890     0.09414
#> 3    -0.00366    -0.03383    -0.08752     0.04668    -0.04189     0.22260
#> 4     0.02689    -0.04810     0.03904    -0.05279     0.07445    -0.05314
#> 5     0.25383    -0.60782    -0.22061    -0.28658     0.25104    -0.35192
#> 6     0.10357     0.00114     0.04931     0.00910    -0.05893     0.05464
#>   90450015805 90248015805 90231015802 90408015802 90290015802 90259015805
#> 1     0.08663    -0.06132     0.04218    -0.00423     0.08077    -0.01780
#> 2     0.25806     0.07516     0.03730     0.11030     0.08670     0.03676
#> 3    -0.00869    -0.06754     0.09629     0.01667    -0.06542     0.03569
#> 4     0.07546    -0.06911     0.05649    -0.02294     0.06344    -0.09505
#> 5    -0.60487    -0.31868    -0.02655    -0.15067    -0.31331    -0.09932
#> 6    -0.04883     0.04513    -0.14700     0.06921     0.00002     0.09950
#>   90422015805 90564015805 90225015805 90581015805
#> 1     0.10560    -0.02055     0.12116    -0.06640
#> 2     0.13345    -0.07851     0.20068    -0.03770
#> 3     0.13243    -0.10413    -0.15158     0.03569
#> 4     0.08954    -0.08688     0.07869    -0.07954
#> 5    -0.10424     0.28558    -0.16343    -0.17998
#> 6    -0.07070     0.01900    -0.01075     0.01471
head(df_to_numeric(df))
#>                90237015805 90245015805 90441015805 90420015805 90289015805
#> XP_017456475.1     0.00769    -0.05952     0.10992    -0.00229     0.04674
#> XP_017447817.1     0.09326    -0.08879     0.31763     0.07739    -0.06031
#> NP_446341.1       -0.11469     0.10623     0.08138     0.08464    -0.16039
#> NP_071796.2        0.05903    -0.06490     0.05673    -0.07494     0.06263
#> NP_001157776.1     1.24616     0.63626     0.15487    -0.39231    -0.64953
#> XP_003754359.1    -0.04140    -0.00964    -0.08768     0.10905    -0.01300
#>                90267015805 90218015805 90585015805 90432015802 90560015805
#> XP_017456475.1     0.06119     0.11000    -0.02633     0.08646    -0.06782
#> XP_017447817.1     0.10648     0.14415     0.12522     0.13868    -0.03150
#> NP_446341.1       -0.09960    -0.09566    -0.02896    -0.11469     0.26114
#> NP_071796.2       -0.02216     0.10700    -0.07468     0.05919    -0.09779
#> NP_001157776.1     0.63982     0.12769    -0.46366    -1.49219    -1.48255
#> XP_003754359.1     0.06778     0.05061     0.03325    -0.14632     0.00766
#>                90430015805 90412015805 90217015805 90266015805 90223015805
#> XP_017456475.1    -0.00225     0.01435     0.07922    -0.01652     0.11231
#> XP_017447817.1     0.09258     0.08732     0.12027    -0.02472     0.16095
#> NP_446341.1        0.14095    -0.03176    -0.16405    -0.13004    -0.07527
#> NP_071796.2        0.08112    -0.06524     0.09477    -0.03734     0.09170
#> NP_001157776.1     0.14749    -0.36679    -0.34689    -0.00593     0.29859
#> XP_003754359.1    -0.04890     0.00656    -0.09305    -0.08272     0.03596
#>                90578015805 90449015805 90258015805 90283015805 90571015805
#> XP_017456475.1    -0.02321     0.09507     0.00787     0.07536    -0.07617
#> XP_017447817.1     0.07759     0.05649     0.19117    -0.02809     0.08864
#> NP_446341.1       -0.01265    -0.14445     0.06341    -0.01666     0.28993
#> NP_071796.2       -0.10477     0.08221    -0.05721     0.05121    -0.12634
#> NP_001157776.1    -0.57429    -0.30849    -0.10142    -0.03489    -0.38285
#> XP_003754359.1    -0.03905    -0.00406     0.10679     0.04905     0.03972
#>                90281015805 90587015805 90222015805 90256015802 90239015805
#> XP_017456475.1     0.07170     0.02030     0.09243    -0.03238     0.04958
#> XP_017447817.1     0.18576     0.09423     0.12652     0.04303     0.06640
#> NP_446341.1       -0.12354    -0.00561    -0.14562     0.05136    -0.01653
#> NP_071796.2        0.06304    -0.01797     0.05310    -0.06212     0.03588
#> NP_001157776.1     0.10696    -0.20935    -0.09217    -0.31013    -0.12688
#> XP_003754359.1     0.03327     0.02120    -0.00913    -0.06928    -0.03088
#>                90559015805 90439015805 90406015805 90423015805 90255015802
#> XP_017456475.1    -0.05368     0.03511    -0.00482     0.07782     0.00998
#> XP_017447817.1     0.08883     0.13454    -0.01024     0.14147    -0.04832
#> NP_446341.1        0.09173     0.04071    -0.08929    -0.03453     0.15072
#> NP_071796.2       -0.06982     0.00694    -0.02401     0.07256    -0.04749
#> NP_001157776.1    -0.10168    -0.06460    -0.53958     0.02972    -0.36775
#> XP_003754359.1    -0.01289    -0.03503     0.00117    -0.03446     0.10633
#>                90227015805 90567015805 90229015805 90576015805 90280015805
#> XP_017456475.1     0.12675    -0.10506     0.07816    -0.04595     0.10060
#> XP_017447817.1     0.26463    -0.00900     0.03422    -0.05347     0.17520
#> NP_446341.1       -0.15487     0.16851    -0.01903    -0.01581    -0.22340
#> NP_071796.2        0.11555    -0.08556     0.04804    -0.07206     0.08251
#> NP_001157776.1     0.32162    -0.01294    -0.06972    -0.66205     0.14655
#> XP_003754359.1     0.07847    -0.03708    -0.07135    -0.00649    -0.00404
#>                90252015805 90445015802 90254015805 90421015805 90416015805
#> XP_017456475.1    -0.04869     0.07865     0.01499     0.06699    -0.00042
#> XP_017447817.1     0.05656     0.24797     0.04646     0.15554    -0.09590
#> NP_446341.1       -0.06180    -0.09393     0.21671     0.05044     0.05260
#> NP_071796.2       -0.06693     0.04238    -0.05073     0.05478    -0.05785
#> NP_001157776.1    -0.41869     0.03080    -0.41675    -0.33470    -0.25958
#> XP_003754359.1    -0.05693    -0.03708     0.08520    -0.03783     0.05743
#>                90426015805 90265015805 90292015805 90574015802 90234015802
#> XP_017456475.1     0.07749    -0.02423     0.11087    -0.06783     0.03175
#> XP_017447817.1     0.16798     0.07771     0.13045    -0.06292     0.15962
#> NP_446341.1        0.05966    -0.06085    -0.17454    -0.00723    -0.00366
#> NP_071796.2        0.08925    -0.06092     0.07279    -0.07737     0.02689
#> NP_001157776.1     0.14447    -0.23248    -0.04627    -0.57912     0.25383
#> XP_003754359.1    -0.05527    -0.06289    -0.04862    -0.02176     0.10357
#>                90410015805 90444015805 90569015802 90232015805 90251015805
#> XP_017456475.1     0.03331     0.05923    -0.01780     0.05311     0.01011
#> XP_017447817.1     0.09036     0.19256    -0.03660     0.00890     0.09414
#> NP_446341.1       -0.03383    -0.08752     0.04668    -0.04189     0.22260
#> NP_071796.2       -0.04810     0.03904    -0.05279     0.07445    -0.05314
#> NP_001157776.1    -0.60782    -0.22061    -0.28658     0.25104    -0.35192
#> XP_003754359.1     0.00114     0.04931     0.00910    -0.05893     0.05464
#>                90450015805 90248015805 90231015802 90408015802 90290015802
#> XP_017456475.1     0.08663    -0.06132     0.04218    -0.00423     0.08077
#> XP_017447817.1     0.25806     0.07516     0.03730     0.11030     0.08670
#> NP_446341.1       -0.00869    -0.06754     0.09629     0.01667    -0.06542
#> NP_071796.2        0.07546    -0.06911     0.05649    -0.02294     0.06344
#> NP_001157776.1    -0.60487    -0.31868    -0.02655    -0.15067    -0.31331
#> XP_003754359.1    -0.04883     0.04513    -0.14700     0.06921     0.00002
#>                90259015805 90422015805 90564015805 90225015805 90581015805
#> XP_017456475.1    -0.01780     0.10560    -0.02055     0.12116    -0.06640
#> XP_017447817.1     0.03676     0.13345    -0.07851     0.20068    -0.03770
#> NP_446341.1        0.03569     0.13243    -0.10413    -0.15158     0.03569
#> NP_071796.2       -0.09505     0.08954    -0.08688     0.07869    -0.07954
#> NP_001157776.1    -0.09932    -0.10424     0.28558    -0.16343    -0.17998
#> XP_003754359.1     0.09950    -0.07070     0.01900    -0.01075     0.01471

df = load_sample_data("SKM-GN", "TRNSCRPT")
#> TRNSCRPT_SKMGN_NORM_DATA
head(df)
#>   feature         feature_ID tissue    assay 90560015512 90581015512
#> 1    <NA> ENSRNOG00000000008 SKM-GN TRNSCRPT     0.04044    -0.09760
#> 2    <NA> ENSRNOG00000000012 SKM-GN TRNSCRPT     2.61487     2.78841
#> 3    <NA> ENSRNOG00000000021 SKM-GN TRNSCRPT     2.17049     1.70091
#> 4    <NA> ENSRNOG00000000024 SKM-GN TRNSCRPT     1.79255     1.26161
#> 5    <NA> ENSRNOG00000000033 SKM-GN TRNSCRPT     5.39651     5.50703
#> 6    <NA> ENSRNOG00000000034 SKM-GN TRNSCRPT     0.60845     0.70086
#>   90406015512 90449015512 90254015512 90283015512 90239015512 90248015512
#> 1    -0.70731     0.13853    -0.44053     0.52784     0.22980    -0.12894
#> 2     2.95085     2.46788     2.44837     2.04327     2.09265     2.32352
#> 3     1.69314     1.90443     2.19652     1.76931     1.86663     1.86467
#> 4     1.67725     1.66870     1.74834     1.90716     1.83755     2.17411
#> 5     5.50525     5.57206     5.37429     5.31720     5.45335     5.55503
#> 6     0.61674     0.33073     0.38421    -0.21000     0.43786     0.70860
#>   90258015512 90578015512 90245015512 90576015512 90281015512 90217015512
#> 1    -0.31359    -0.02966     0.52824    -0.81585     0.34030     0.40735
#> 2     2.10696     2.11677     2.20265     2.54491     2.19650     2.37995
#> 3     1.95060     2.00223     1.84537     1.64870     1.84650     1.95022
#> 4     1.73023     1.20399     1.48707     1.45265     1.40510     1.69679
#> 5     5.32968     5.63894     5.54827     5.42096     5.53512     5.40322
#> 6     0.64727    -0.10104     0.00451     0.30108    -0.07392    -0.01300
#>   90218015512 90229015512 90267015512 90587015512 90559015512 90232015512
#> 1     0.03677     0.31649    -1.24165    -0.73293    -0.02849    -0.48280
#> 2     2.77938     2.06254     2.46108     2.66164     2.62302     2.21977
#> 3     1.93337     2.11891     1.71588     1.92069     1.78806     2.08880
#> 4     1.77839     1.71007     1.69699     1.69943     1.58249     1.38897
#> 5     5.60805     5.58965     5.69878     5.61886     5.68709     5.47694
#> 6     0.58318    -0.54629    -0.30874     0.10947     0.38316    -0.01816
#>   90289015512 90420015512 90421015512 90585015512 90422015512 90412015512
#> 1    -0.12810     0.18212     0.21443     0.03242     0.72636    -0.53412
#> 2     2.82647     1.95962     2.68798     2.47048     2.79476     2.48789
#> 3     1.97686     1.60349     1.79380     1.96145     2.01976     1.73055
#> 4     1.63852     1.66007     1.46482     1.73534     1.32286     1.66532
#> 5     5.59153     5.29348     5.52934     5.49770     5.52196     5.60471
#> 6     0.19214     0.50402     0.11879     0.15948     0.12390    -0.28424
#>   90430015512 90237015512 90225015512 90410015512 90265015512 90252015512
#> 1     0.41563    -0.26578     0.40170    -0.31662    -0.62897    -0.17535
#> 2     2.25362     2.41558     2.77491     2.43681     2.16175     2.72017
#> 3     1.86286     1.82707     1.94617     1.85790     2.03610     1.77483
#> 4     1.74606     1.51519     1.55420     1.56966     1.60364     1.25378
#> 5     5.25659     5.59528     5.63889     5.52594     5.38533     5.45564
#> 6    -0.30661    -0.02640     0.08141    -0.73109     0.37754     0.19093
#>   90227015512 90426015512 90439015512 90450015512 90423015512 90416015512
#> 1    -0.11433     0.12451     0.15498     0.13228     0.57511     0.05581
#> 2     2.69406     2.84558     2.13412     2.18922     2.66813     2.55399
#> 3     2.04226     1.81523     1.82251     1.76915     2.09234     1.87355
#> 4     1.72742     1.41653     1.24493     1.76797     1.53528     1.71780
#> 5     5.68308     5.38394     5.57047     5.48786     5.46659     5.40784
#> 6    -0.00864     0.48256    -0.32330     0.34597     0.87286     0.27035
#>   90292015512 90567015512 90441015512 90222015512 90444015512 90251015512
#> 1     0.81147    -1.46482     0.26997    -0.11452    -0.58334     0.01797
#> 2     2.02226     2.42018     2.88176     2.39418     2.22883     2.52931
#> 3     1.30966     2.21349     1.68039     1.82031     1.97314     1.98996
#> 4     1.38656     1.61739     1.71741     1.55077     1.66902     1.54092
#> 5     5.54606     5.60948     5.47114     5.72685     5.62799     5.53311
#> 6    -0.31298    -0.05936     0.31897     0.44893     0.74196     0.32654
#>   90571015512 90564015512 90223015512 90259015512 90280015512 90266015512
#> 1    -0.41208    -0.33233     0.04380    -0.73042    -0.85329     0.19635
#> 2     2.39657     2.23037     2.42393     2.86850     2.46706     2.49215
#> 3     1.90911     0.96063     1.86491     1.84083     1.62856     1.73748
#> 4     1.77951     1.24067     1.50975     1.57042     1.94865     1.71153
#> 5     5.58606     5.69502     5.60403     5.40084     5.42323     5.28813
#> 6     0.31401    -1.27254     0.15498     0.64676     0.16397     0.62764
head(df_to_numeric(df))
#>                    90560015512 90581015512 90406015512 90449015512 90254015512
#> ENSRNOG00000000008     0.04044    -0.09760    -0.70731     0.13853    -0.44053
#> ENSRNOG00000000012     2.61487     2.78841     2.95085     2.46788     2.44837
#> ENSRNOG00000000021     2.17049     1.70091     1.69314     1.90443     2.19652
#> ENSRNOG00000000024     1.79255     1.26161     1.67725     1.66870     1.74834
#> ENSRNOG00000000033     5.39651     5.50703     5.50525     5.57206     5.37429
#> ENSRNOG00000000034     0.60845     0.70086     0.61674     0.33073     0.38421
#>                    90283015512 90239015512 90248015512 90258015512 90578015512
#> ENSRNOG00000000008     0.52784     0.22980    -0.12894    -0.31359    -0.02966
#> ENSRNOG00000000012     2.04327     2.09265     2.32352     2.10696     2.11677
#> ENSRNOG00000000021     1.76931     1.86663     1.86467     1.95060     2.00223
#> ENSRNOG00000000024     1.90716     1.83755     2.17411     1.73023     1.20399
#> ENSRNOG00000000033     5.31720     5.45335     5.55503     5.32968     5.63894
#> ENSRNOG00000000034    -0.21000     0.43786     0.70860     0.64727    -0.10104
#>                    90245015512 90576015512 90281015512 90217015512 90218015512
#> ENSRNOG00000000008     0.52824    -0.81585     0.34030     0.40735     0.03677
#> ENSRNOG00000000012     2.20265     2.54491     2.19650     2.37995     2.77938
#> ENSRNOG00000000021     1.84537     1.64870     1.84650     1.95022     1.93337
#> ENSRNOG00000000024     1.48707     1.45265     1.40510     1.69679     1.77839
#> ENSRNOG00000000033     5.54827     5.42096     5.53512     5.40322     5.60805
#> ENSRNOG00000000034     0.00451     0.30108    -0.07392    -0.01300     0.58318
#>                    90229015512 90267015512 90587015512 90559015512 90232015512
#> ENSRNOG00000000008     0.31649    -1.24165    -0.73293    -0.02849    -0.48280
#> ENSRNOG00000000012     2.06254     2.46108     2.66164     2.62302     2.21977
#> ENSRNOG00000000021     2.11891     1.71588     1.92069     1.78806     2.08880
#> ENSRNOG00000000024     1.71007     1.69699     1.69943     1.58249     1.38897
#> ENSRNOG00000000033     5.58965     5.69878     5.61886     5.68709     5.47694
#> ENSRNOG00000000034    -0.54629    -0.30874     0.10947     0.38316    -0.01816
#>                    90289015512 90420015512 90421015512 90585015512 90422015512
#> ENSRNOG00000000008    -0.12810     0.18212     0.21443     0.03242     0.72636
#> ENSRNOG00000000012     2.82647     1.95962     2.68798     2.47048     2.79476
#> ENSRNOG00000000021     1.97686     1.60349     1.79380     1.96145     2.01976
#> ENSRNOG00000000024     1.63852     1.66007     1.46482     1.73534     1.32286
#> ENSRNOG00000000033     5.59153     5.29348     5.52934     5.49770     5.52196
#> ENSRNOG00000000034     0.19214     0.50402     0.11879     0.15948     0.12390
#>                    90412015512 90430015512 90237015512 90225015512 90410015512
#> ENSRNOG00000000008    -0.53412     0.41563    -0.26578     0.40170    -0.31662
#> ENSRNOG00000000012     2.48789     2.25362     2.41558     2.77491     2.43681
#> ENSRNOG00000000021     1.73055     1.86286     1.82707     1.94617     1.85790
#> ENSRNOG00000000024     1.66532     1.74606     1.51519     1.55420     1.56966
#> ENSRNOG00000000033     5.60471     5.25659     5.59528     5.63889     5.52594
#> ENSRNOG00000000034    -0.28424    -0.30661    -0.02640     0.08141    -0.73109
#>                    90265015512 90252015512 90227015512 90426015512 90439015512
#> ENSRNOG00000000008    -0.62897    -0.17535    -0.11433     0.12451     0.15498
#> ENSRNOG00000000012     2.16175     2.72017     2.69406     2.84558     2.13412
#> ENSRNOG00000000021     2.03610     1.77483     2.04226     1.81523     1.82251
#> ENSRNOG00000000024     1.60364     1.25378     1.72742     1.41653     1.24493
#> ENSRNOG00000000033     5.38533     5.45564     5.68308     5.38394     5.57047
#> ENSRNOG00000000034     0.37754     0.19093    -0.00864     0.48256    -0.32330
#>                    90450015512 90423015512 90416015512 90292015512 90567015512
#> ENSRNOG00000000008     0.13228     0.57511     0.05581     0.81147    -1.46482
#> ENSRNOG00000000012     2.18922     2.66813     2.55399     2.02226     2.42018
#> ENSRNOG00000000021     1.76915     2.09234     1.87355     1.30966     2.21349
#> ENSRNOG00000000024     1.76797     1.53528     1.71780     1.38656     1.61739
#> ENSRNOG00000000033     5.48786     5.46659     5.40784     5.54606     5.60948
#> ENSRNOG00000000034     0.34597     0.87286     0.27035    -0.31298    -0.05936
#>                    90441015512 90222015512 90444015512 90251015512 90571015512
#> ENSRNOG00000000008     0.26997    -0.11452    -0.58334     0.01797    -0.41208
#> ENSRNOG00000000012     2.88176     2.39418     2.22883     2.52931     2.39657
#> ENSRNOG00000000021     1.68039     1.82031     1.97314     1.98996     1.90911
#> ENSRNOG00000000024     1.71741     1.55077     1.66902     1.54092     1.77951
#> ENSRNOG00000000033     5.47114     5.72685     5.62799     5.53311     5.58606
#> ENSRNOG00000000034     0.31897     0.44893     0.74196     0.32654     0.31401
#>                    90564015512 90223015512 90259015512 90280015512 90266015512
#> ENSRNOG00000000008    -0.33233     0.04380    -0.73042    -0.85329     0.19635
#> ENSRNOG00000000012     2.23037     2.42393     2.86850     2.46706     2.49215
#> ENSRNOG00000000021     0.96063     1.86491     1.84083     1.62856     1.73748
#> ENSRNOG00000000024     1.24067     1.50975     1.57042     1.94865     1.71153
#> ENSRNOG00000000033     5.69502     5.60403     5.40084     5.42323     5.28813
#> ENSRNOG00000000034    -1.27254     0.15498     0.64676     0.16397     0.62764

df = MotrpacRatTraining6moData::METAB_NORM_DATA_FLAT 
head(df)
#>                                   feature                 feature_ID tissue
#> 1                                    <NA>          1-Methylhistidine SKM-GN
#> 2          METAB;SKM-GN;3-Methylhistidine          3-Methylhistidine SKM-GN
#> 3                                    <NA>                    Alanine SKM-GN
#> 4 METAB;SKM-GN;alpha-Amino-N-butyric-acid alpha-Amino-N-butyric-acid SKM-GN
#> 5     METAB;SKM-GN;alpha-Aminoadipic-acid     alpha-Aminoadipic-acid SKM-GN
#> 6                                    <NA>                   Anserine SKM-GN
#>   assay        dataset  10023259  10024735  10025626  10025707  10025979
#> 1 METAB metab-t-amines -2.021833 -2.079719 -3.436179 -2.892081 -2.887504
#> 2 METAB metab-t-amines -5.701379 -5.434643 -3.936401 -5.480303 -5.793896
#> 3 METAB metab-t-amines  3.187501  3.215838  2.607263  3.012177  3.231277
#> 4 METAB metab-t-amines -4.377759 -4.316324 -5.241371 -4.899344 -5.146494
#> 5 METAB metab-t-amines -7.332883 -7.265484 -7.700365 -7.354299 -7.479045
#> 6 METAB metab-t-amines  5.338719  5.366602  4.814032  5.075998  5.400620
#>    10026193  10026355  10026789  10027327  10027599  10043527  10043950
#> 1 -4.514741 -2.544423 -3.375448 -3.939732 -2.543649 -3.742979 -3.513945
#> 2 -5.507136 -5.778351 -5.901492 -5.853154 -5.965013 -5.831052 -5.831052
#> 3  2.784447  3.492982  2.871950  3.036416  2.767701  2.976358  2.984420
#> 4 -5.864239 -3.842643 -4.678553 -5.150986 -5.322288 -2.204102 -2.749800
#> 5 -7.903082 -7.455229 -7.667897 -7.458262 -8.063798 -7.204498 -7.611380
#> 6  4.876223  5.642179  5.249762  5.333676  5.034198  4.975659  5.104240
#>    10044256  10044337  10044760  10045228  10045309  10046119  10046380
#> 1 -3.270733 -4.636199 -4.094534 -5.253733 -4.747603 -5.619634 -3.713415
#> 2 -6.322980 -5.465335 -5.517073 -5.961423 -5.568810 -5.878088 -5.697338
#> 3  2.899921  2.709388  2.664008  3.363926  2.955020  3.034935  2.428681
#> 4 -2.836623 -2.286044 -2.914355 -4.410404 -5.024094 -5.357220 -3.009668
#> 5 -7.763101 -7.539778 -7.471515 -7.574276 -7.599955 -7.606123 -7.430794
#> 6  4.858562  4.884447  4.602911  4.926568  4.892441  4.960555  4.782479
#>    10046461  10095241  10106383  10139044  10185585  10220917   10306471
#> 1 -5.572583 -2.543268 -4.279139 -2.046517 -3.888372 -3.536587  -4.020920
#> 2 -5.732703 -5.793896 -6.163963 -5.691029 -5.878088 -5.853154  -6.103870
#> 3  2.933419  3.186209  2.873747  3.048724  3.026919  3.044117   2.337171
#> 4 -4.370965 -4.376463 -5.513140 -3.983984 -4.225540 -4.665510  -4.292606
#> 5 -7.581218 -7.245138 -7.469203 -7.412318 -7.249989 -7.314580 -10.742948
#> 6  4.781347  5.353303  5.131960  5.521528  5.116090  5.221344   4.450209
#>    10335064   10337199  10381414   10411690   10417702  10422463   10424911
#> 1 -4.497613  -4.183032 -3.802254  -4.702437  -4.291564 -2.434529  -3.095332
#> 2 -6.103870  -6.212347 -6.707517  -6.103870  -3.936401 -5.895734  -6.103870
#> 3  2.393124   3.196304  2.990990   2.583516   2.683816  3.260951   2.929117
#> 4 -3.782016  -4.605558 -3.986041  -4.503596  -5.011978 -4.300478  -4.568007
#> 5 -9.849630 -10.511145 -9.188837 -10.570954 -10.080625 -9.594837 -10.872973
#> 6  4.452848   4.836833  5.007872   4.438323   4.937824  5.763951   5.177845
#>     10503110   10561072  10486003   10502300  10587543  10641858  10672656
#> 1  -2.970098  -2.754089 -2.488646  -3.127141 -3.283126 -3.005427 -2.565457
#> 2  -6.212347  -6.212347 -5.982330  -6.212347 -6.351455 -6.052962 -6.013115
#> 3   2.625345   2.917452  3.147056   2.929198  2.828940  3.240191  3.153487
#> 4  -4.880179  -4.725287 -3.748631  -5.488511 -3.825740 -4.646943 -4.314681
#> 5 -10.719532 -10.821435 -9.325438 -11.889597 -9.504741 -9.353949 -9.436099
#> 6   4.990016   5.293500  5.656533   5.284779  5.172667  5.697340  5.598126
#>    10950486  10700102  10833331  10901434  10953744  10729518  10734945
#> 1 -4.325997 -4.314607 -3.533223 -3.451202 -3.939811 -3.755320 -3.784486
#> 2 -6.056082 -5.995394 -6.001171 -5.592065 -6.043406 -6.796087 -5.448060
#> 3  2.805928  2.772517  2.569808  3.308268  2.805006  3.024528  2.831024
#> 4 -2.700609 -3.218246 -2.393703 -2.151927 -3.051718 -1.959642 -2.846545
#> 5 -9.065578 -9.990355 -9.432046 -8.109909 -9.338194 -7.884048 -9.165749
#> 6  4.575264  5.036168  4.852671  5.466784  5.037051  5.189719  5.161469
#>    10761160  10871349   10971645 10044841 10044922 10204989 10314733
#> 1 -3.333975 -3.576662  -3.616002       NA       NA       NA       NA
#> 2 -5.853714 -5.977130  -6.090829       NA       NA       NA       NA
#> 3  2.858938  2.743712   2.502620       NA       NA       NA       NA
#> 4 -2.187656 -2.728209  -2.693416       NA       NA       NA       NA
#> 5 -8.530027 -8.327523 -10.083005       NA       NA       NA       NA
#> 6  5.279703  5.287191   4.925035       NA       NA       NA       NA
rn = paste(df$assay, df$tissue, df$feature_ID, df$dataset, sep=";")
head(df_to_numeric(df, rownames = rn))
#>                                                         10023259  10024735
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -2.021833 -2.079719
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.701379 -5.434643
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.187501  3.215838
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -4.377759 -4.316324
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.332883 -7.265484
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.338719  5.366602
#>                                                         10025626  10025707
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.436179 -2.892081
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -3.936401 -5.480303
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.607263  3.012177
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -5.241371 -4.899344
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.700365 -7.354299
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.814032  5.075998
#>                                                         10025979  10026193
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -2.887504 -4.514741
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.793896 -5.507136
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.231277  2.784447
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -5.146494 -5.864239
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.479045 -7.903082
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.400620  4.876223
#>                                                         10026355  10026789
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -2.544423 -3.375448
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.778351 -5.901492
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.492982  2.871950
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.842643 -4.678553
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.455229 -7.667897
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.642179  5.249762
#>                                                         10027327  10027599
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.939732 -2.543649
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.853154 -5.965013
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.036416  2.767701
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -5.150986 -5.322288
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.458262 -8.063798
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.333676  5.034198
#>                                                         10043527  10043950
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.742979 -3.513945
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.831052 -5.831052
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.976358  2.984420
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.204102 -2.749800
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.204498 -7.611380
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.975659  5.104240
#>                                                         10044256  10044337
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.270733 -4.636199
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.322980 -5.465335
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.899921  2.709388
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.836623 -2.286044
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.763101 -7.539778
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.858562  4.884447
#>                                                         10044760  10045228
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -4.094534 -5.253733
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.517073 -5.961423
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.664008  3.363926
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.914355 -4.410404
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.471515 -7.574276
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.602911  4.926568
#>                                                         10045309  10046119
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -4.747603 -5.619634
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.568810 -5.878088
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.955020  3.034935
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -5.024094 -5.357220
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.599955 -7.606123
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.892441  4.960555
#>                                                         10046380  10046461
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.713415 -5.572583
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.697338 -5.732703
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.428681  2.933419
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.009668 -4.370965
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.430794 -7.581218
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.782479  4.781347
#>                                                         10095241  10106383
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -2.543268 -4.279139
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.793896 -6.163963
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.186209  2.873747
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -4.376463 -5.513140
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.245138 -7.469203
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.353303  5.131960
#>                                                         10139044  10185585
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -2.046517 -3.888372
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.691029 -5.878088
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.048724  3.026919
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.983984 -4.225540
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.412318 -7.249989
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.521528  5.116090
#>                                                         10220917   10306471
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.536587  -4.020920
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.853154  -6.103870
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.044117   2.337171
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -4.665510  -4.292606
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -7.314580 -10.742948
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.221344   4.450209
#>                                                         10335064   10337199
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -4.497613  -4.183032
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.103870  -6.212347
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.393124   3.196304
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.782016  -4.605558
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.849630 -10.511145
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.452848   4.836833
#>                                                         10381414   10411690
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.802254  -4.702437
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.707517  -6.103870
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.990990   2.583516
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.986041  -4.503596
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.188837 -10.570954
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.007872   4.438323
#>                                                          10417702  10422463
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines           -4.291564 -2.434529
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines           -3.936401 -5.895734
#> METAB;SKM-GN;Alanine;metab-t-amines                      2.683816  3.260951
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines  -5.011978 -4.300478
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -10.080625 -9.594837
#> METAB;SKM-GN;Anserine;metab-t-amines                     4.937824  5.763951
#>                                                          10424911   10503110
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines           -3.095332  -2.970098
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines           -6.103870  -6.212347
#> METAB;SKM-GN;Alanine;metab-t-amines                      2.929117   2.625345
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines  -4.568007  -4.880179
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -10.872973 -10.719532
#> METAB;SKM-GN;Anserine;metab-t-amines                     5.177845   4.990016
#>                                                          10561072  10486003
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines           -2.754089 -2.488646
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines           -6.212347 -5.982330
#> METAB;SKM-GN;Alanine;metab-t-amines                      2.917452  3.147056
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines  -4.725287 -3.748631
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -10.821435 -9.325438
#> METAB;SKM-GN;Anserine;metab-t-amines                     5.293500  5.656533
#>                                                          10502300  10587543
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines           -3.127141 -3.283126
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines           -6.212347 -6.351455
#> METAB;SKM-GN;Alanine;metab-t-amines                      2.929198  2.828940
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines  -5.488511 -3.825740
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -11.889597 -9.504741
#> METAB;SKM-GN;Anserine;metab-t-amines                     5.284779  5.172667
#>                                                         10641858  10672656
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.005427 -2.565457
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.052962 -6.013115
#> METAB;SKM-GN;Alanine;metab-t-amines                     3.240191  3.153487
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -4.646943 -4.314681
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.353949 -9.436099
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.697340  5.598126
#>                                                         10950486  10700102
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -4.325997 -4.314607
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.056082 -5.995394
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.805928  2.772517
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.700609 -3.218246
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.065578 -9.990355
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.575264  5.036168
#>                                                         10833331  10901434
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.533223 -3.451202
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.001171 -5.592065
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.569808  3.308268
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.393703 -2.151927
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.432046 -8.109909
#> METAB;SKM-GN;Anserine;metab-t-amines                    4.852671  5.466784
#>                                                         10953744  10729518
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.939811 -3.755320
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -6.043406 -6.796087
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.805006  3.024528
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -3.051718 -1.959642
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.338194 -7.884048
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.037051  5.189719
#>                                                         10734945  10761160
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.784486 -3.333975
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.448060 -5.853714
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.831024  2.858938
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.846545 -2.187656
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -9.165749 -8.530027
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.161469  5.279703
#>                                                         10871349   10971645
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines          -3.576662  -3.616002
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines          -5.977130  -6.090829
#> METAB;SKM-GN;Alanine;metab-t-amines                     2.743712   2.502620
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines -2.728209  -2.693416
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines     -8.327523 -10.083005
#> METAB;SKM-GN;Anserine;metab-t-amines                    5.287191   4.925035
#>                                                        10044841 10044922
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines                NA       NA
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines                NA       NA
#> METAB;SKM-GN;Alanine;metab-t-amines                          NA       NA
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines       NA       NA
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines           NA       NA
#> METAB;SKM-GN;Anserine;metab-t-amines                         NA       NA
#>                                                        10204989 10314733
#> METAB;SKM-GN;1-Methylhistidine;metab-t-amines                NA       NA
#> METAB;SKM-GN;3-Methylhistidine;metab-t-amines                NA       NA
#> METAB;SKM-GN;Alanine;metab-t-amines                          NA       NA
#> METAB;SKM-GN;alpha-Amino-N-butyric-acid;metab-t-amines       NA       NA
#> METAB;SKM-GN;alpha-Aminoadipic-acid;metab-t-amines           NA       NA
#> METAB;SKM-GN;Anserine;metab-t-amines                         NA       NA