An auxiliary function for getting the top analyte sets for a set of tissues.

An auxiliary function for getting the top analyte sets for a set of tissues.

Usage

extract_tissue_sets(
  tissues,
  node_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$node_sets,
  edge_sets = MotrpacRatTraining6moData::GRAPH_COMPONENTS$edge_sets,
  k = 3,
  min_size = 20,
  add_week8 = TRUE,
  omes = NULL
)

Arguments

tissues

A character vector. The names of the tissues (one or more) to be considered.

node_sets

A named list with the node (state) sets of analytes/features. GRAPH_COMPONENTS$node_sets by default.

edge_sets

A named list with the edge (state) sets of analytes/features. GRAPH_COMPONENTS$edge_sets by default.

k

An integer. How many node/edge/trajectory sets to extract. See details.

min_size

An integer specifying the minimal path size to be considered.

add_week8

A logical. TRUE (default): add all week 8 nodes to the node set.

omes

A character vector (optional). The names of the omes (one or more) to be considered.

Value

A named list. Names correspond to set names (can correspond to a node, edge, or a trajectory).

Details

This function is useful for extracting the largest sets for a specific set of tissues and omes. Thus, it is a useful step before running enrichment analysis on sets of analytes identified by the graphical clustering analysis.

By specifying k the user can control how many sets to include. The default is three, which means that the top largest node sets, edge sets, and full trajectories will be extracted (three each).

Naming format: Analyte names are in the ome;tissue;feature_id format. Node set names: week_Fdifferential regulation status_Mdifferential regulation status. Example 1: 1w_F1_M0 means up-regulation in females week 1 and null (zero effect) in males in week 1. Example 2: 1w_F1_M-1 means up-regulation in females week 1 and down-regulation in males in week 1. The edge set object is a named list of string vectors. The name of an edge is node_id---node_id. A full trajectory is annotated by node_id->node_id->node_id->node_id. For example: 1w_F-1_M-1->2w_F-1_M-1->4w_F-1_M-1->8w_F-1_M-1.

See bayesian_graphical_clustering for more details about the graphical analysis.

Examples

filtered_solution = extract_tissue_sets(
   tissues = c("HEART","SKM-VL","SKM-GN")
)